QuantumControl

Documentation for QuantumControl.

Control problems

QuantumControlBase.ControlProblem — Type

A full control problem with multiple objectives.

ControlProblem(
   objectives=<list of objectives>,
   pulse_options=<dict of controls to pulse options>,
   tlist=<time grid>,
   kwargs...
)

Note that the control problem can only be instantiated via keyword arguments.

The objectives are a list of Objective instances, each defining an initial state, a dynamical generator for the evolution of the state, and (optionally) a target for the evolution.

The pulse_options are a dictionary (IdDict) mapping controls that occur in the objectives to properties specific to the control method.

The tlist is the time grid on which the time evolution of the initial states of each objective should be propagated.

The remaining kwargs are keyword arguments that are passed directl to the optimal control method. These typically include e.g. the optimization functional.

The control problem is solved by finding a set of controls that simultaneously fulfill all objectives.

QuantumControlBase.Objective — Type

A single optimization objective.

Objective(
    initial_state=<intial state>,
    generator=<dynamical generator>,
    [target=<optional target state or specification>]
)

Note that the objective can only be instantiated via keyword arguments.

Discretization

QuantumControlBase.discretize — Function

Evaluate control at every point of tlist.

values = discretize(control, tlist; via_midpoints=true)

discretizes the given control to a Vector of values defined on the points of tlist.

If control is a function, it will will first be evaluated at the midpoint of tlist, see discretize_on_midpoints, and then the values on the midpoints are converted to values on tlist. This discretization is more stable than directly evaluationg the control function at the values of tlist, and ensures that repeated round-trips between discretize and discretize_on_midpoints can be done safely, see the note in the documentation of discretize_on_midpoints.

The latter can still be achieved by passing via_midpoints=false. While such a direct discretization is suitable e.g. for plotting, but it is unsuitable for round-trips between discretize and discretize_on_midpoints (constant controls on tlist may result in a zig-zag on the intervals of tlist).

If control is a vector, it will be returned un-modified if it is of the same length as tlist. Otherwise, control must have one less value than tlist, and is assumed to be defined on the midpoins of tlist. In that case, discretize acts as the inverse of discretize_on_midpoints. See discretize_on_midpoints for how control values on tlist and control values on the intervals of tlist are related.

QuantumControlBase.discretize_on_midpoints — Function

Evaluate control at the midpoints of tlist.

values = discretize_on_midpoints(control, tlist)

discretizes the given control to a Vector of values on the midpoints of tlist. Hence, the resulting values will contain one less value than tlist.

If control is a vector of values defined on tlist (i.e., of the same length as tlist), it will be converted to a vector of values on the intervals of tlist. The value for the first and last "midpoint" will remain the original values at the beginning and end of tlist, in order to ensure exact bounary conditions. For all other midpoints, the value for that midpoint will be calculated by "un-averaging".

For example, for a control and tlist of length 5, consider the following diagram:

tlist index:       1   2   3   4   5
tlist:             ⋅   ⋅   ⋅   ⋅   ⋅   input values cᵢ (i ∈ 1..5)
                   |̂/ ̄ ̄ ̂\ / ̂\ / ̂ ̄ ̄\|̂
midpoints:         x     x   x     x   output values pᵢ (i ∈ 1..4)
midpoints index:   1     2   3     4

We will have $p₁=c₁$ for the first value, $p₄=c₅$ for the last value. For all other points, the control values $cᵢ = \frac{p_{i-1} + p_{i}}{2}$ are the average of the values on the midpoints. This implies the "un-averaging" for the midpoint values $pᵢ = 2 c_{i} - p_{i-1}$.

Note

An arbitrary input control array may not be compatible with the above averaging formula. In this case, the conversion will be "lossy" (discretize will not recover the original control array; the difference should be considered a "discretization error"). However, any further round-trip conversions between points and intervals are bijective and preserve the boundary conditions. In this case, the discretize_on_midpoints and discretize methods are each other's inverse. This also implies that for an optimal control procedure, it is safe to modify midpoint values. Modifying the the values on the time grid directly on the other hand may accumulate discretization errors.

If control is a vector of one less length than tlist, it will be returned unchanged, under the assumption that the input is already properly discretized.

If control is a function, the function will be directly evaluated at the midpoints marked as x in the above diagram..

Time dependencies

QuantumControlBase.setcontrolvals — Function

Construct G by plugging values into a general generator.

G = setcontrolvals(generator, vals_dict)
setcontrolvals!(G, generator, vals_dict)

evaluates the specific dynamical generator G by plugging in values into the general generator according to vals_dict.

The vals_dict is a dictionary (IdDict) mapping controls as returned by getcontrols(generator) to values.

QuantumControlBase.setcontrolvals! — Function

In-place version of setcontrolvals.

QuantumControlBase.getcontrols — Function

Extract a Tuple of controls.

controls = getcontrols(generator)

extracts the controls from a single dynamical generator.

controls = getcontrols(objectives)

extracts the controls from a list of objectives (i.e., from each objective's generator)

In either case, controls that occur multiple times, either in a single generator, or throughout the different objectives, will occur only once in the result.

By default, assumes that any generator is a nested Tuple, e.g. (H0, (H1, ϵ1), (H2, ϵ2), ...) and extracts (ϵ1, ϵ2)

Each control must be a valid argument for discretize.

Control shapes

QuantumControlBase.flattop — Function

Flat shape (one) with a switch-on/switch-off from zero.

flattop(t; t₀, T, t_rise, t_fall=t_rise, func=:blackman)

evaluates a shape function that starts at 0 at $t=t₀$, and ramps to to 1 during the t_rise interval. The function then remains at value 1, before ramping down to 0 again during the interval t_fall before T. For $t < t₀$ and $t > T$, the shape is zero.

The default switch-on/-off shape is half of a Blackman window (see blackman).

For func=:sinsq, the switch-on/-off shape is a sine-squared curve.

QuantumControlBase.box — Function

Box shape (Theta-function).

box(t, t₀, T)

evaluates the Heaviside (Theta-) function $\Theta(t) = 1$ for $t_0 \le t \le T$; and $\Theta(t) = 0$ otherwise.

QuantumControlBase.blackman — Function

Blackman window shape.

blackman(t, t₀, T; a=0.16)

calculates

\[B(t; t_0, T) = \frac{1}{2}\left( 1 - a - \cos\left(2π \frac{t - t_0}{T - t_0}\right) + a \cos\left(4π \frac{t - t_0}{T - t_0}\right) \right)\,,\]

for a scalar t, with $a$ = 0.16.

See http://en.wikipedia.org/wiki/Window_function#Blackman_windows

A Blackman shape looks nearly identical to a Gaussian with a 6-sigma interval between t₀ and T. Unlike the Gaussian, however, it will go exactly to zero at the edges. Thus, Blackman pulses are often preferable to Gaussians.

Propagation

QuantumPropagators.propagate — Function

Propagate a state over an entire time grid.

propagate(state, genfunc, tlist;
          backwards=false; wrk=nothing, method=:auto, storage=nothing,
          observables=(<store state>, ), hook=nothing)

propagates state over the time grid in tlist, using piecewise-constant dynamical generators (Hamiltonians or Liouvillians) determined by genfunc, and returns the resulting propagated state. The propagation is performed by calling propstep! for every interval in tlist.

For the i'th time interval, genfunc(tlist, i) must return the generator for that time interval. Generally, when approximating a time-continuous dynamical generator as piecewise-constant on the time grid, it should be evaluated at the midpoint of the interval. A possible exception is the first and last interval, which may be better evaluated at tlist[1] and tlist[end] to ensure exact boundary conditions like control fields that are exactly zero.

In addition to the two positional parameters indicating the time interval, genfunc will also receive the state (the input state for the propagation step), backwards, storage, observables, and init as keyword arguments. These additional parameters may be used for unusual equations of motion beyond the standard Schrödinger or Liouville-von-Neumann equation, e.g. state would enter the genfunc for a Gross–Pitaevskii equation. For standard equations of motion that do not use the additional parameters, it is best to capture the keyword arguments to genfunc with a definition like

genfunc(tlist, i; kwargs...) = ...

The propagation method is determined by wrk, see initpropwrk and `propstep!: If wrk is not given, it will be created internally for the given method with initpropwrk; for the default method=:auto, initpropwrk will attempt to choose the most appropriate method for the given parameters, using the generator returned by genfunc with i=1 and the keyword argument init=true.

In general, there is no requirement that tlist has a constant time step, although some propagation methods (most notably cheby!) only support a uniform time grid.

If storage is given as an Array, it will be filled with data determined by the observables. The default "observable" results in the propagated states at every point in time being stored. The storage array should be created with init_storage. See its documentation for details.

The storage parameter may also be given as true, and a new storage array will be created internally with init_storage and returned instead of the propagated state.

If backwards is true, the input state is assumed to be at time tlist[end], and the propagation progresses backwards in time (with a negative time step dt). If storage is given, it will be filled back-to-front during the backwards propagation.

If hook is given as a callable, it will be called after each propagation step, as hook(state, generator, tlist, i, wrk, observables) where i is the index of the time interval on tlist covered by the propagation step (0 for the initial state, respectives lastindex(tlist) for the backward propagation). The hook is called before calculating any observables. Example usage includes writing data to file, or modyfing state, e.g. removing amplitude from the lowest and highest level to mitigate "truncation error".

The propagate routine returns the propagated state at tlist[end], respectively tlist[1] if backwards=true, or a storage array with the stored states / observable data if storage=true.

Propagate the initial state of a control objective.

propagate(obj, tlist; controls_map=IdDict(), kwargs...)

propagates obj.initial_state under the dynamics described by obj.generator.

The optional dict control_map may be given to replace the controls in obj.generator (as obtained by getcontrols) with custom functions or vectors, e.g. with the controls resulting from optimization.

All kwargs are forwarded to QuantumPropagators.propagate.

QuantumPropagators.initpropwrk — Function

Initialize a workspace for propagation.

wrk = initpropwrk(state, tlist, generator...; method=:auto, kwargs...)

The resulting wrk can be passed to propagate or propstep!.

Arguments

state: An exemplary state for the propagation (e.g., the initial state)

tlist: The time grid over which propagate will be called. Must include at least to points in order to determine the propagation time step to prepare. If the propagation will be over a tlist with a variable dt, the full tlist must be passed here.

generator: An exemplary (non-time-dependent) dynamical generator. For full generality (if method=:cheby), the given generator should have a spectral range sufficiently large to encompass the entire propagation. If given multiple times, a spectral envelope enclosing all the generators will be determined automatically. In this case, you should pass the generators with the extremal values of all the controls.

method: The propagation method to use. The default value of :auto attempts to choose the best method available, based on the properties of the given state, tlist, and generator. Alternative values are :cheby and :newton, and :expprop.

All other kwargs are filtered and passed to the contructor for returned workspace, e.g. limit for method=:cheby or m_max for method=:newton.

wrk = initpropwrk(obj, tlist; method=:auto, kwargs...)

initializes a workspace for the propagation of a control Objective.

QuantumPropagators.propstep! — Function

Perform a single propagation step in-place.

propstep!(state, generator, dt, wrk;, kwargs...)

The propagation method is determined by wrk, see initpropwrk. The kwargs are forwarded to the underlying method

QuantumPropagators.ChebyWrk — Type

Workspace for the Chebychev propagation routine.

    ChebyWrk(Ψ, Δ, E_min, dt; limit=1e-12)

initializes the workspace for the propagation of a state similar to Ψ under a Hamiltonian with eigenvalues between E_min and E_min + Δ, and a time step dt. Chebychev coefficients smaller than the given limit are discarded.

QuantumPropagators.cheby_coeffs! — Function

Calculate Chebychev coefficients in-place.

Writes the coefficients into coeffs. Returns the number n of coefficients required for convergence, or the size of coeffs, whichever is less.

coeffs: An array to hold the Chebychev cofficients. On output, the elements 1 through n will hold the calculated coefficients, while the remaining elements will be unchanged

dt: the time step

QuantumPropagators.cheby_coeffs — Function

Calculate Chebychev coefficients.

Return an array of coefficiencts larger than limit.

Args:

Δ: the spectral radius of the underlying operator

dt: the time step

QuantumPropagators.NewtonWrk — Type

    NewtonWrk(v0, m_max=10)

Workspace for the Newton-with-restarted-Arnoldi propagation routine.

Initializes the workspace for the propagation of a vector v0, using a maximum Krylov dimension of m_max in each restart iteration. Note that m_max should be smaller than the length of v0.

QuantumPropagators.ExpPropWrk — Type

    ExpPropWrk(v0)

Workspace for propagation via direct matrix exponentiation.

Initializes the workspace for the propagation of a vector v0

QuantumPropagators.cheby! — Function

Evaluate Ψ = exp(-i H dt) Ψ in-place.

Args:

Ψ: on input, initial vector. Will be overwritten with result.

H: Hermitian operator

dt: time step

E_min: minimum eigenvalue of H, to be used instead of the E_min from the initialization of wrk. The same wrk may be used for different values E_min, as long as the spectra radius Δ and the time step dt are the same as those used for the initialization of wrk.

The routine will not allocate any internal storage. This implementation requires copyto! lmul!, and axpy! to be implemented for Ψ, and the three-argument mul! for Ψ and H.

QuantumPropagators.newton! — Function

newton!(Ψ, H, dt, wrk, func=(z -> exp(-1im*z)); norm_min=1e-14, relerr=1e-12,
        max_restarts=50)

Evaluate Ψ = func(H*dt) Ψ using a Newton-with-restarted-Arnoldi scheme.

Arguments

Ψ: The state to propagate, will be overwritten in-place with the propagated state

H: Operator acting on Ψ. Together with dt, this is the argument to func

dt: Implicit time step. Together with H, this is the argument to func

wkr: Work array, initialized with NewtonWrk

func: The function to apply to H dt, taking a single (scalar) complex-valued argument z in place of H dt. The default func is to evaluate the time evoluation operator for the Schrödinger equation

norm_min: the minium norm at which to consider a state similar to Ψ as zero

relerr: The relative error defining the convergence condition for the restart iteration. Propagation stops when the norm of the accumulated Ψ is stable up to the given relative error

max_restart: The maximum number of restart iterations. Exceeding max_restart will throw an AssertionError.

QuantumPropagators.expprop! — Function

expprop!(Ψ, H, dt, wrk, func=(z -> exp(-1im*z)))

Evaluate Ψ = func(H*dt) Ψ by directly evaluating U = func(H*dt), i.e. by matrix exponentiation for the default func, and then multiplying U and Ψ in-place with mul!.

The workspace wrk must be initialized with ExpPropWrk to provide storage for a temporary state.

Storage

QuantumPropagators.init_storage — Function

Create a storage array for propagate.

storage = init_storage(state, tlist)

creates a storage array suitable for storing a state for each point in tlist.

storage = init_storage(state, tlist, observables))

creates a storage array suitable for the data generated by the observables applied to state, see map_observables, for each point in tlist.

storage = init_storage(data, nt))

creates a storage arrays suitable for storing data nt times, where nt=length(tlist). By default, this will be a vector of typeof(data) and length nt, or a n × nt Matrix with the same eltype as data if data is a Vector of length n.

QuantumPropagators.write_to_storage! — Function

Place data into storage for time slot i.

    write_to_storage!(storage, i, state, observables)

For a storage array created by init_storage, store the data obtains from map_observables into the storage for time slot i. This delegates to the more general

    write_to_storage!(storage, i, data)

Conceptually, this corresponds roughly to storage[i] = data, but storage may have its own idea on how to store data for a specific time slot. For example, with the default init_storage Vector data will be stored in a matrix, and write_to_storage! will in this case write data to the i'th column of the matrix.

For a given type of storage and data, it is the developer's responsibility that init_storage and write_to_storage! are compatible.

QuantumPropagators.get_from_storage! — Function

Obtain data from storage

get_from_storage!(state, storage, i)

extracts data from the storage for the i'th time slot. Invese of write_to_storage!

QuantumPropagators.map_observable — Function

Apply a single observable to state.

data = map_observable(observable, state)

By default, observable is assumed to be callable, and the above is equivalent to data = observable(state).

If observable is a matrix and state is a vector evaluate the expectation value of the observable as dot(state, observable, state).

QuantumPropagators.map_observables — Function

Obtain "observable" data from state.

data = map_observables(observables, state)

calculates the data for a tuple of observables applied to state. For a single observable (tuple of length 1), simply return the result of map_observable.

For multiple observables, return the tuple resulting from applying map_observable for each observable. If the tuple is "uniform" (all elements are of the same type, e.g. if each observable calculates the expectation value of a Hermitian operator), it is converted to a Vector. This allows for compact storage in a storage array, see init_storage.

Optimization

Krotov.optimize_pulses — Function

Use Krotov's method to optimize the given optimization problem.

result = optimize_pulses(problem)

optimizes the control problem, see QuantumControlBase.ControlProblem.

Parameters are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

The optimization functional is given implicitly via the mandatory problem keyword argument chi.

Optional problem keyword arguments

The following problem keyword arguments are supported (with default values):

sigma=nothing: Function that calculate the second-order contribution. If not given, the first-order Krotov method is used.

iter_start=0: the initial iteration number

iter_stop=5000: the maximum iteration number

check_convergence: a function to check whether convergence has been reached. Receives a KrotovResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘.

prop_method=:auto: The propagation method to use

Index

QuantumControlBase.ControlProblem

QuantumControlBase.Objective

QuantumPropagators.ChebyWrk

QuantumPropagators.ExpPropWrk

QuantumPropagators.NewtonWrk

Krotov.optimize_pulses

QuantumControlBase.blackman

QuantumControlBase.box

QuantumControlBase.discretize

QuantumControlBase.discretize_on_midpoints

QuantumControlBase.flattop

QuantumControlBase.getcontrols

QuantumControlBase.setcontrolvals

QuantumControlBase.setcontrolvals!

QuantumPropagators.cheby!

QuantumPropagators.cheby_coeffs

QuantumPropagators.cheby_coeffs!

QuantumPropagators.expprop!

QuantumPropagators.get_from_storage!

QuantumPropagators.init_storage

QuantumPropagators.initpropwrk

QuantumPropagators.map_observable

QuantumPropagators.map_observables

QuantumPropagators.newton!

QuantumPropagators.propagate

QuantumPropagators.propstep!

QuantumPropagators.write_to_storage!