API

API

QuantumControl

QuantumControl.shapes

QuantumControl.functionals

Krotov

Public

Private

QuantumControlBase

Public

Private

QuantumPropagators

Public

Private

Index

QuantumControl

QuantumControl.optimize — Function

Optimize a quantum control problem.

opt_result = optimize(problem; method=<method>, kwargs...)

optimizes towards a solution of given problem with the given optimization method. All keyword arguments update (overwrite) parameters in problem

opt_result = optimize(problem; method=:krotov, kwargs...)

optimizes problem using Krotov's method, see Krotov.optimize_pulses.

All other types and methods in QuantumControl or its submodules are re-exported from lower level packages (QuantumControlBase, QuantumPropagators, etc.):

ControlProblem

Objective

WeightedObjective

liouvillian

propagate

propstep!

QuantumControl.shapes

QuantumControl.shapes.flattop

QuantumControl.shapes.box

QuantumControl.shapes.blackman

QuantumControl.functionals

QuantumControl.functionals.F_ss

QuantumControl.functionals.J_T_ss

QuantumControl.functionals.chi_ss!

QuantumControl.functionals.F_sm

QuantumControl.functionals.J_T_sm

QuantumControl.functionals.chi_sm!

QuantumControl.functionals.F_re

QuantumControl.functionals.J_T_re

QuantumControl.functionals.chi_re!

Krotov

Public

Krotov.SquareParametrization — Function

Parametrization ϵ(t) = u²(t), enforcing pulse values ≥ 0.

Krotov.TanhParametrization — Function

Parametrization with a tanh function that enforces ϵmin < ϵ(t) < ϵmax.

Krotov.TanhSqParametrization — Function

Parametrization with a tanh² function that enforces $0 ≤ ϵ(t) < ϵ_max$.

Missing docstring.

Missing docstring for Krotov.LogisticParametrization, Krotov.LogisticSqParametrization. Check Documenter's build log for details.

Krotov.optimize_pulses — Function

Use Krotov's method to optimize the given optimization problem.

result = optimize_pulses(problem; kwargs...)

optimizes the given control problem, see QuantumControlBase.ControlProblem.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem. Any kwargs passed directly to optimize_pulses will update (overwrite) the parameters in problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives) that evaluates the final time functional from a list ϕ of forward-propagated states and problem.objectives.

chi: A function chi!(χ, ϕ, objectives) what receives a list ϕ of the forward propagates state and must set $χₖ=∂J_T/∂⟨ϕₖ|$.

Optional problem keyword arguments

The following keyword arguments are supported (with default values):

sigma=nothing: Function that calculate the second-order contribution. If not given, the first-order Krotov method is used.

iter_start=0: the initial iteration number

iter_stop=5000: the maximum iteration number

prop_method=:auto: The propagation method to use

update_hook: A function that receives the Krotov workspace, the iteration number, the list of updated pulses and the list of guess pulses as positional arguments. The function may mutate any of its arguments. This may be used e.g. to apply a spectral filter to the updated pulses, or to update propagation workspaces inside the Krotov workspace.

info_hook: A function that receives the same argumens as update_hook, in order to write information about the current iteration to the screen or to a file. The default info_hook prints a table with convergence information to the screen. Runs after update_hook. The info_hook function may return a tuple, which is stored in the list of records inside the KrotovResult object.

check_convergence: a function to check whether convergence has been reached. Receives a KrotovResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘. The convergence check is performed after any calls to update_hook and info_hook.

Private

Krotov.KrotovResult — Type

Result object returned by optimize_pulses.

Krotov.PulseParametrization — Type

Specification for a "time-local" pulse parametrization.

The parametrization is given as a collection of three functions:

$ϵ(u(t))$

$u(ϵ(t))$

$∂ϵ/∂u$ as a function of $u(t)$.

Krotov.NoParametrization — Method

Parametrization where $ϵ(t) ≡ u(t)$.

Krotov.print_table — Method

Default info_hook

QuantumControlBase

Public

QuantumControlBase.ControlProblem — Type

A full control problem with multiple objectives.

ControlProblem(
   objectives=<list of objectives>,
   pulse_options=<dict of controls to pulse options>,
   tlist=<time grid>,
   kwargs...
)

Note that the control problem can only be instantiated via keyword arguments.

The objectives are a list of AbstractControlObjective instances, each defining an initial state and a dynamical generator for the evolution of that state. Usually, the objective will also include a target state (see Objective) and possibly a weight (see WeightedObjective.

The pulse_options are a dictionary (IdDict) mapping controls that occur in the objectives to properties specific to the control method.

The tlist is the time grid on which the time evolution of the initial states of each objective should be propagated.

The remaining kwargs are keyword arguments that are passed directly to the optimal control method. These typically include e.g. the optimization functional.

The control problem is solved by finding a set of controls that simultaneously fulfill all objectives.

QuantumControlBase.Objective — Type

Standard optimization objective.

Objective(;
    initial_state=<initial_state>,
    generator=<genenerator>,
    target_state=<target_state>
)

describes an optimization objective where the time evoluation of the given initial_state under the given generator aims towards target_state. The generator here is e.g. a time-dependent Hamiltonian or Liouvillian.

The most common control problems in quantum control, e.g. state-to-state transitions or quantum gate implementations can be expressed by simultaneously fulfilling multiple objectives of this type.

Note that the objective can only be instantiated via keyword arguments.

QuantumControlBase.WeightedObjective — Type

Standard optimization objective with a weight.

WeightedObjective(;
    initial_state=<initial_state>,
    generator=<genenerator>,
    target_state=<target_state>,
    weight=<weight>
)

initializes a control objective like Objective, but with an additional weight parameter (a float generally between 0 and 1) that weights the objective relative to other objectives that are part of the same control problem.

QuantumControlBase.F_re — Method

Real-part fidelity.

F_re(ϕ, objectives; τ=nothing)

QuantumControlBase.F_sm — Method

Square-modulus fidelity.

F_sm(ϕ, objectives; τ=nothing)

QuantumControlBase.F_ss — Method

State-to-state phase-insensitive fidelity.

F_ss(ϕ, objectives; τ=nothing)

QuantumControlBase.J_T_re — Method

Real-part functional.

J_T_re(ϕ, objectives; τ=nothing)

QuantumControlBase.J_T_sm — Method

Square-modulus functional.

J_T_sm(ϕ, objectives; τ=nothing)

QuantumControlBase.J_T_ss — Method

State-to-state phase-insensitive functional.

J_T_ss(ϕ, objectives; τ=nothing)

QuantumControlBase.blackman — Method

Blackman window shape.

blackman(t, t₀, T; a=0.16)

calculates

\[B(t; t_0, T) = \frac{1}{2}\left( 1 - a - \cos\left(2π \frac{t - t_0}{T - t_0}\right) + a \cos\left(4π \frac{t - t_0}{T - t_0}\right) \right)\,,\]

for a scalar t, with $a$ = 0.16.

See http://en.wikipedia.org/wiki/Window_function#Blackman_windows

A Blackman shape looks nearly identical to a Gaussian with a 6-sigma interval between t₀ and T. Unlike the Gaussian, however, it will go exactly to zero at the edges. Thus, Blackman pulses are often preferable to Gaussians.

QuantumControlBase.box — Method

Box shape (Theta-function).

box(t, t₀, T)

evaluates the Heaviside (Theta-) function $\Theta(t) = 1$ for $t_0 \le t \le T$; and $\Theta(t) = 0$ otherwise.

QuantumControlBase.chi_re! — Method

Krotov-states χ for functional J_T_re.

chi_re!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.chi_sm! — Method

Krotov-states χ for functional J_T_sm.

chi_sm!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.chi_ss! — Method

Krotov-states χ for functional J_T_ss.

chi_ss!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.discretize — Method

Evaluate control at every point of tlist.

values = discretize(control, tlist; via_midpoints=true)

discretizes the given control to a Vector of values defined on the points of tlist.

If control is a function, it will will first be evaluated at the midpoint of tlist, see discretize_on_midpoints, and then the values on the midpoints are converted to values on tlist. This discretization is more stable than directly evaluationg the control function at the values of tlist, and ensures that repeated round-trips between discretize and discretize_on_midpoints can be done safely, see the note in the documentation of discretize_on_midpoints.

The latter can still be achieved by passing via_midpoints=false. While such a direct discretization is suitable e.g. for plotting, but it is unsuitable for round-trips between discretize and discretize_on_midpoints (constant controls on tlist may result in a zig-zag on the intervals of tlist).

If control is a vector, it will be returned un-modified if it is of the same length as tlist. Otherwise, control must have one less value than tlist, and is assumed to be defined on the midpoins of tlist. In that case, discretize acts as the inverse of discretize_on_midpoints. See discretize_on_midpoints for how control values on tlist and control values on the intervals of tlist are related.

QuantumControlBase.discretize_on_midpoints — Method

Evaluate control at the midpoints of tlist.

values = discretize_on_midpoints(control, tlist)

discretizes the given control to a Vector of values on the midpoints of tlist. Hence, the resulting values will contain one less value than tlist.

If control is a vector of values defined on tlist (i.e., of the same length as tlist), it will be converted to a vector of values on the intervals of tlist. The value for the first and last "midpoint" will remain the original values at the beginning and end of tlist, in order to ensure exact bounary conditions. For all other midpoints, the value for that midpoint will be calculated by "un-averaging".

For example, for a control and tlist of length 5, consider the following diagram:

tlist index:       1   2   3   4   5
tlist:             ⋅   ⋅   ⋅   ⋅   ⋅   input values cᵢ (i ∈ 1..5)
                   |̂/ ̄ ̄ ̂\ / ̂\ / ̂ ̄ ̄\|̂
midpoints:         x     x   x     x   output values pᵢ (i ∈ 1..4)
midpoints index:   1     2   3     4

We will have $p₁=c₁$ for the first value, $p₄=c₅$ for the last value. For all other points, the control values $cᵢ = \frac{p_{i-1} + p_{i}}{2}$ are the average of the values on the midpoints. This implies the "un-averaging" for the midpoint values $pᵢ = 2 c_{i} - p_{i-1}$.

Note

An arbitrary input control array may not be compatible with the above averaging formula. In this case, the conversion will be "lossy" (discretize will not recover the original control array; the difference should be considered a "discretization error"). However, any further round-trip conversions between points and intervals are bijective and preserve the boundary conditions. In this case, the discretize_on_midpoints and discretize methods are each other's inverse. This also implies that for an optimal control procedure, it is safe to modify midpoint values. Modifying the the values on the time grid directly on the other hand may accumulate discretization errors.

If control is a vector of one less length than tlist, it will be returned unchanged, under the assumption that the input is already properly discretized.

If control is a function, the function will be directly evaluated at the midpoints marked as x in the above diagram..

QuantumControlBase.flattop — Method

Flat shape (one) with a switch-on/switch-off from zero.

flattop(t; t₀, T, t_rise, t_fall=t_rise, func=:blackman)

evaluates a shape function that starts at 0 at $t=t₀$, and ramps to to 1 during the t_rise interval. The function then remains at value 1, before ramping down to 0 again during the interval t_fall before T. For $t < t₀$ and $t > T$, the shape is zero.

The default switch-on/-off shape is half of a Blackman window (see blackman).

For func=:sinsq, the switch-on/-off shape is a sine-squared curve.

QuantumControlBase.getcontrols — Method

Extract a Tuple of controls.

controls = getcontrols(generator)

extracts the controls from a single dynamical generator.

controls = getcontrols(objectives)

extracts the controls from a list of objectives (i.e., from each objective's generator)

In either case, controls that occur multiple times, either in a single generator, or throughout the different objectives, will occur only once in the result.

By default, assumes that any generator is a nested Tuple, e.g. (H0, (H1, ϵ1), (H2, ϵ2), ...) and extracts (ϵ1, ϵ2)

Each control must be a valid argument for discretize.

QuantumControlBase.liouvillian — Function

Construct a Liouvillian super-operator.

ℒ = liouvillian(Ĥ, c_ops=(); convention=:LvN)

calculates the sparse Liouvillian super-operator ℒ from the Hamiltonian Ĥ and a list c_ops of Lindblad operators.

With convention=:LvN, applying the resulting ℒ to a vectorized density matrix ρ⃗ calculates $\frac{d}{dt} \vec{\rho}(t) = ℒ \vec{\rho}(t)$ equivalent to the Liouville-von-Neumann equation for the density matrix $ρ̂$,

\[\frac{d}{dt} ρ̂(t) = -i [Ĥ, ρ̂(t)] + \sum_k\left( Â_k ρ̂ Â_k^\dagger - \frac{1}{2} A_k^\dagger Â_k ρ̂ - \frac{1}{2} ρ̂ Â_k^\dagger Â_k \right)\,,\]

where the Lindblad operators $Â_k$ are the elements of c_ops.

The Hamiltonian $Ĥ$ may be time-dependent, using a nested-tuple format by default, e.g., (Ĥ₀, (H₁, ϵ₁), (H₂, ϵ₂)), where ϵ₁ and ϵ₂ are functions of time. In this case, the resulting ℒ will also be in nested tuple format, ℒ = (ℒ₀, (ℒ₁, ϵ₁), (ℒ₂, ϵ₂)), where the initial element contains the superoperator ℒ₀ for the static component of the Liouvillian, i.e., the commutator with the drift Hamiltonian Ĥ₀, plus the dissipator (sum over $k$), as a sparse matrix. Time-dependent Lindblad operators are not supported. The remaining elements are tuples (ℒ₁, ϵ₁) and (ℒ₂, ϵ₂) corresponding to the commutators with the two control Hamiltonians, where ℒ₁ and ℒ₂ again are sparse matrices.

If $Ĥ$ is not time-dependent, the resulting ℒ will be a single-element tuple containing the Liouvillian as a sparse matrix, ℒ = (ℒ₀, ).

With convention=:TDSE, the Liouvillian will be constructed for the equation of motion $-i \hbar \frac{d}{dt} \vec{\rho}(t) = ℒ \vec{\rho}(t)$ to match exactly the form of the time-dependent Schrödinger equation. While this notation is not standard in the literature of open quantum systems, it has the benefit that the resulting ℒ can be used in a numerical propagator for a (non-Hermitian) Schrödinger equation without any change. Thus, for numerical applications, convention=:TDSE is generally preferred. The returned ℒ between the two conventions differs only by a factor of $i$, since we generally assume $\hbar=1$.

The convention keyword argument is mandatory, to force a conscious choice.

See Goerz et. al. "Optimal control theory for a unitary operation under dissipative evolution", arXiv 1312.0111v2, Appendix B.2 for the explicit construction of the Liouvillian superoperator as a sparse matrix.

QuantumControlBase.obj_genfunc — Method

Construct a genfunc suitable for propagate from an objective.

genfunc = obj_genfunc(obj, tlist; controls_map=IdDict())

can be passed to propagate to propagate under the dynamical generator in obj. If given, controls_map defines replacements for the control

If given, control_map defines replacements for the controls in obj.generator. This allows, e.g., to replace the controls with those resulting from an optimization.

QuantumControlBase.setcontrolvals! — Method

In-place version of setcontrolvals.

QuantumControlBase.setcontrolvals — Method

Construct G by plugging values into a general generator.

G = setcontrolvals(generator, vals_dict)
setcontrolvals!(G, generator, vals_dict)

evaluates the specific dynamical generator G by plugging in values into the general generator according to vals_dict.

The vals_dict is a dictionary (IdDict) mapping controls as returned by getcontrols(generator) to values.

Private

QuantumControlBase.AbstractControlObjective — Type

Base class for a single optimization objective.

All objectives must have a field initial_state and a field generator, at minimum.

Base.adjoint — Function

adjoint(objective)

Adjoint of a control objective. The adjoint objective contains the adjoint of the dynamical generator obj.generator, and adjoints of the obj.initial_state / obj.target_state if these exist and have an adjoint.

QuantumControlBase.f_tau — Function

Average complex overlap of the target states with forward-propagated states.

f_tau(ϕ, objectives; τ=nothing)

QuantumControlBase.initobjpropwrk — Function

wrk = initobjpropwrk(obj, tlist, method; kwargs...)

initializes a workspace for the propagation of anAbstractControlObjective.

QuantumPropagators

Public

QuantumPropagators.specrange — Function

Calculate the spectral range of a Hamiltonian H on the real axis.

E_min, E_max = specrange(H; method=:auto, kwargs...)

calculates the approximate lowest and highest eigenvalues of H. Any imaginary part in the eigenvalues is ignored: the routine is intended for (although not strictly limited to) a Hermitian H.

This delegates to

specrange(H, method; kwargs...)

for the different methods.

The default method=:auto chooses the best method for the given H. This is :diag for small matrices, and :arnoldi otherwise. Keyword arguments not relevant to the underlying implementation will be ignored.

E_min, E_max = specrange(H, :arnoldi; state=random_state(H), m_min=20,
                         m_max=60, prec=1e-3, nom_min=1e-15, enlarge=true)

uses Arnoldi iteration with state as the starting vector. It approximates the eigenvalues of H with between m_min and m_max Ritz values, until the lowest and highest eigenvalue are stable to a releative precision of prec. The norm_min parameter is passed to the underlying arnoldi!.

If enlarge=true (default) the returned E_min and E_max will be enlarged via a heuristic to slightly over-estimate the spectral radius instead of under-estimating it.

E_min, E_max = specrange(H, :diag)

uses exact diagonization via the standard eigenvals function to obtain the smallest and largest eigenvalue. This should only be used for relatively small matrices.

QuantumPropagators.cheby_coeffs — Function

Calculate Chebychev coefficients.

Return an array of coefficiencts larger than limit.

Arguments

Δ: the spectral radius of the underlying operator

dt: the time step

QuantumPropagators.cheby_coeffs! — Function

Calculate Chebychev coefficients in-place.

n = cheby_coeffs!(coeffs, Δ, dt, limit=1e-12)

overwrites the first n values in coeffs with new coefficients larger than limit for the given new spectral radius Δ and time step dt. The coeffs array will be resized if necessary, and may length > n on exit.

QuantumPropagators.ChebyWrk — Type

Workspace for the Chebychev propagation routine.

    ChebyWrk(Ψ, Δ, E_min, dt; limit=1e-12)

initializes the workspace for the propagation of a state similar to Ψ under a Hamiltonian with eigenvalues between E_min and E_min + Δ, and a time step dt. Chebychev coefficients smaller than the given limit are discarded.

QuantumPropagators.cheby! — Function

Evaluate Ψ = exp(-i H dt) Ψ in-place.

cheby!(Ψ, H, dt, wrk; E_min=nothing, check_normalization=false)

Args:

Ψ: on input, initial vector. Will be overwritten with result.

H: Hermitian operator

dt: time step

wrk: internal workspace

E_min: minimum eigenvalue of H, to be used instead of the E_min from the initialization of wrk. The same wrk may be used for different values E_min, as long as the spectra radius Δ and the time step dt are the same as those used for the initialization of wrk.

check_normalizataion: perform checks that the H does not exceed the spectral radius for which the the workspace was initialized.

The routine will not allocate any internal storage. This implementation requires copyto! lmul!, and axpy! to be implemented for Ψ, and the three-argument mul! for Ψ and H.

QuantumPropagators.NewtonWrk — Type

    NewtonWrk(v0, m_max=10)

Workspace for the Newton-with-restarted-Arnoldi propagation routine.

Initializes the workspace for the propagation of a vector v0, using a maximum Krylov dimension of m_max in each restart iteration. Note that m_max should be smaller than the length of v0.

QuantumPropagators.newton! — Function

newton!(Ψ, H, dt, wrk; func=(z -> exp(-1im*z)), norm_min=1e-14, relerr=1e-12,
        max_restarts=50)

Evaluate Ψ = func(H*dt) Ψ using a Newton-with-restarted-Arnoldi scheme.

Arguments

Ψ: The state to propagate, will be overwritten in-place with the propagated state

H: Operator acting on Ψ. Together with dt, this is the argument to func

dt: Implicit time step. Together with H, this is the argument to func

wkr: Work array, initialized with NewtonWrk

func: The function to apply to H dt, taking a single (scalar) complex-valued argument z in place of H dt. The default func is to evaluate the time evoluation operator for the Schrödinger equation

norm_min: the minium norm at which to consider a state similar to Ψ as zero

relerr: The relative error defining the convergence condition for the restart iteration. Propagation stops when the norm of the accumulated Ψ is stable up to the given relative error

max_restart: The maximum number of restart iterations. Exceeding max_restart will throw an AssertionError.

QuantumPropagators.ExpPropWrk — Type

    ExpPropWrk(v0)

Workspace for propagation via direct matrix exponentiation.

Initializes the workspace for the propagation of a vector v0

QuantumPropagators.expprop! — Function

expprop!(Ψ, H, dt, wrk; func=(H_dt -> exp(-1im * H_dt)))

Evaluate Ψ = func(H*dt) Ψ by directly evaluating U = func(H*dt), i.e. by matrix exponentiation for the default func, and then multiplying U and Ψ in-place with mul!.

The workspace wrk must be initialized with ExpPropWrk to provide storage for a temporary state.

QuantumPropagators.init_storage — Function

Create a storage array for propagate.

storage = init_storage(state, tlist)

creates a storage array suitable for storing a state for each point in tlist.

storage = init_storage(state, tlist, observables))

creates a storage array suitable for the data generated by the observables applied to state, see map_observables, for each point in tlist.

storage = init_storage(data, nt))

creates a storage arrays suitable for storing data nt times, where nt=length(tlist). By default, this will be a vector of typeof(data) and length nt, or a n × nt Matrix with the same eltype as data if data is a Vector of length n.

QuantumPropagators.map_observables — Function

Obtain "observable" data from state.

data = map_observables(observables, state)

calculates the data for a tuple of observables applied to state. For a single observable (tuple of length 1), simply return the result of map_observable.

For multiple observables, return the tuple resulting from applying map_observable for each observable. If the tuple is "uniform" (all elements are of the same type, e.g. if each observable calculates the expectation value of a Hermitian operator), it is converted to a Vector. This allows for compact storage in a storage array, see init_storage.

QuantumPropagators.map_observable — Function

Apply a single observable to state.

data = map_observable(observable, state)

By default, observable is assumed to be callable, and the above is equivalent to data = observable(state).

If observable is a matrix and state is a vector evaluate the expectation value of the observable as dot(state, observable, state).

QuantumPropagators.write_to_storage! — Function

Place data into storage for time slot i.

    write_to_storage!(storage, i, state, observables)

For a storage array created by init_storage, store the data obtains from map_observables into the storage for time slot i. This delegates to the more general

    write_to_storage!(storage, i, data)

Conceptually, this corresponds roughly to storage[i] = data, but storage may have its own idea on how to store data for a specific time slot. For example, with the default init_storage Vector data will be stored in a matrix, and write_to_storage! will in this case write data to the i'th column of the matrix.

For a given type of storage and data, it is the developer's responsibility that init_storage and write_to_storage! are compatible.

QuantumPropagators.get_from_storage! — Function

Obtain data from storage

get_from_storage!(state, storage, i)

extracts data from the storage for the i'th time slot. Invese of write_to_storage!

QuantumPropagators.initpropwrk — Function

Initialize a workspace for propagation.

wrk = initpropwrk(state, tlist, method=:auto, generator...; kwargs...)

The resulting wrk can be passed to propagate or propstep!.

Arguments

state: An exemplary state for the propagation (e.g., the initial state)

tlist: The time grid over which propagate will be called. Must include at least to points in order to determine the propagation time step to prepare. If the propagation will be over a tlist with a variable dt, the full tlist must be passed here.

generator: An exemplary (non-time-dependent) dynamical generator. For full generality (if method=:cheby), the given generator should have a spectral range sufficiently large to encompass the entire propagation. If given multiple times, a spectral envelope enclosing all the generators will be determined automatically. In this case, you should pass the generators with the extremal values of all the controls.

method: The propagation method to use. The default value of :auto attempts to choose the best method available, based on the properties of the given state, tlist, and generator. Alternative values are :cheby and :newton, and :expprop.

specrad_method: for method=:cheby, method to use for estimating the spectral radius, see specrange. Defaults to :auto.

tolerance: for method=:cheby, a tolerance factor for the estimated spectral radius. That is, Chebychev coefficients will be calculated for a spectral radius increased by the tolerance factor compared to the specral radius estimated for the generator.

All other kwargs are filtered and passed to the contructor for returned workspace, e.g. limit for method=:cheby or m_max for method=:newton. For method=:cheby, they additionally passed to specrange.

QuantumPropagators.propstep! — Function

Perform a single propagation step in-place.

propstep!(state, generator, dt, wrk;, kwargs...)

The propagation method is determined by wrk, see initpropwrk. The kwargs are forwarded to the underlying method

QuantumPropagators.propagate — Function

Propagate a state over an entire time grid.

state_out = propagate(
    state, genfunc, tlist; method=:auto,
    backwards=false; storage=nothing, observables=(<store state>, ),
    hook=nothing, showprogress=false, kwargs...)

propagates state over the time grid in tlist, using piecewise-constant dynamical generators (Hamiltonians or Liouvillians) determined by genfunc, and returns the resulting propagated state. The propagation is performed by calling propstep! for every interval in tlist.

For the i'th time interval, genfunc(tlist, i) must return the generator for that time interval. Generally, when approximating a time-continuous dynamical generator as piecewise-constant on the time grid, it should be evaluated at the midpoint of the interval. A possible exception is the first and last interval, which may be better evaluated at tlist[1] and tlist[end] to ensure exact boundary conditions like control fields that are exactly zero.

In addition to the two positional parameters indicating the time interval, genfunc will also receive the state (the input state for the propagation step), backwards, storage, observables, and init as keyword arguments. These additional parameters may be used for unusual equations of motion beyond the standard Schrödinger or Liouville-von-Neumann equation, e.g. state would enter the genfunc for a Gross–Pitaevskii equation. For standard equations of motion that do not use the additional parameters, it is best to capture the keyword arguments to genfunc with a definition like

genfunc(tlist, i; kwargs...) = ...

For valid propagation methods, see initpropwrk.

In general, there is no requirement that tlist has a constant time step, although some propagation methods (most notably cheby!) only support a uniform time grid.

If storage is given as an Array, it will be filled with data determined by the observables. The default "observable" results in the propagated states at every point in time being stored. The storage array should be created with init_storage. See its documentation for details.

The storage parameter may also be given as true, and a new storage array will be created internally with init_storage and returned instead of the propagated state:

data = propagate(
    state, genfunc, tlist; method=:auto
    backwards=false; storage=true, observables=observables,
    hook=nothing, showprogress=false, kwargs...)

If backwards is true, the input state is assumed to be at time tlist[end], and the propagation progresses backwards in time (with a negative time step dt). If storage is given, it will be filled back-to-front during the backwards propagation.

If hook is given as a callable, it will be called after each propagation step, as hook(state, generator, tlist, i, wrk, observables) where i is the index of the time interval on tlist covered by the propagation step (0 for the initial state, respectives lastindex(tlist) for the backward propagation). The hook is called before calculating any observables. Example usage includes writing data to file, or modyfing state, e.g. removing amplitude from the lowest and highest level to mitigate "truncation error".

If showprogress is given as true, a progress bar will be shown for long-running propagationn. In order to customize the progress bar, showprogress may also be a function that receives length(tlist) and returns a ProgressMeter.Progress instance.

The propagate routine returns the propagated state at tlist[end], respectively tlist[1] if backwards=true, or a storage array with the stored states / observable data if storage=true.

Propagate with the dynamical generator of a control objective.

propagate(obj, tlist; method=:auto, initial_state=obj.initial_state,
          controls_map=IdDict(), kwargs...)

propagates initial_state under the dynamics described by obj.generator.

The optional dict control_map may be given to replace the controls in obj.generator (as obtained by getcontrols) with custom functions or vectors, e.g. with the controls resulting from optimization.

All other kwargs are forwarded to the underlying propagate method for obj.initial_state.

Private

QuantumPropagators.arnoldi! — Method

m = arnoldi!(Hess, q, m, Ψ, H, dt=1.0; extended=true, norm_min=1e-15)

Calculate the Hessenberg matrix and Arnoldi vectors of H dt, from Ψ.

For a given order m, the m×m Hessemberg matrix is calculated and stored in in the pre-allocated Hess. Further an array of m normalized Arnoldi vectors is stored in in the pre-allocated q, plus one additional unnormalized Arnoldi vector. The unnormalized m+1st vector could be used to easily extend a given m×m Hessenberg matrix to a (m+1)×(m+1) matrix.

If the extended Hessenberg matrix is requested (extended=true, default), the m+1st Arnoldi vector is also normalized, and it's norm will be stored in m+1, m entry of the (extended) Hessenberg matrix, which is an (m+1)×(m+1) matrix.

Return the size m of the calculated Hessenberg matrix. This will usually be the input m, except when the Krylov dimension of H starting from Ψ is less then m. E.g., if Ψ is an eigenstate of H, the returned m will be 1.

See http://en.wikipedia.org/wiki/Arnoldi_iteration for a description of the algorithm.

Arguments

Hess::Matrix{ComplexF64}: Pre-allocated storage for the Hessemberg matrix. Can be uninitialized on input. The matrix must be at least of size m×m, or (m+1)×(m+1) if extended=true. On output, the m×m sub-matrix of Hess (with the returned output m) will contain the Hessenberg matrix, and all other elements of Hess be be set to zero.

q: Pre-allocated array of states similar to Ψ, as storage for the calculated Arnoldi vectors. These may be un-initialized on input. Must be at least of length m+1

m: The requested dimensions of the output Hessenberg matrix.

Ψ: The starting vector for the Arnoldi procedure. This can be of any type, as long as Φ = H * Ψ results in a vector similar to Ψ, there is an inner products of Φ and Ψ (Ψ⋅Φ is defined), and norm(Ψ) is defined.

H: The operator (up to dt) for which to calculate the Arnoldi procedure. Can be of any type, as long as H * Ψ is defined.

dt: The implicit time step; the total operator for which to calculate the Arnoldi procedure is H * dt

extended: If true (default), calculate the extended Hessenberg matrix, and normalized the final Arnoldi vector

norm_min: the minimum value of the norm of Ψ at which Ψ should be considered the zero vector

QuantumPropagators.diagonalize_hessenberg_matrix — Method

diagonalize_hessenberg_matrix(Hess, m; accumulate=false)

Diagonalize the m × m top left submatrix of the given Hessenberg matrix.

If accumulate is true, return the concatenated eigenvalues for Hess[1:1,1:1] to Hess[1:m,1:m], that is, all sumatrices of size 1 through m.

QuantumPropagators.extend_arnoldi! — Function

Extend dimension of Hessenberg matrix by one.

extend_arnoldi!(Hess, q, m, H, dt; norm_min=1e-15)

extends the entries in Hess from size (m-1)×(m-1) to size m×m, and the list q of Arnoldi vectors from m to (m+1). It is assumed that the input Hess was created by a call to arnoldi! with extended=false or a previous call to extend_arnoldi!. Note that Hess itself is not resized, so it must be allocated to size m×m or greater on input.

QuantumPropagators.extend_leja! — Method

extend_leja!(leja, n, newpoints, n_use)

Given an array of n (ordered) Leja points, extract n_use points from newpoints, and append them to the existing Leja points. The array leja should be sufficiently large to hold the new Leja points, which are appended after index n_old. It will be re-allocated if necessary and may have a size of up to 2*(n+n_use).

Arguments

leja: Array of leja values. Must contain the "old" leja values to be kept in leja(0:n-1). On output, n_use new leja points will be in leja(n+:n+n_use-1), for the original value of n. The leja array must use zero-based indexing.

n: On input, number of "old" leja points in leja. On output, total number of leja points (i.e. n=n+n_use)

newpoints: On input, candidate points for new leja points. The n_use best values will be chosen and added to leja. On output, the values of new_points are undefined.

n_use: Number of points that should be added to leja

QuantumPropagators.extend_newton_coeffs! — Method

extend_newton_coeffs!(a, n_a, leja, func, n_leja, radius)

Extend the array a of existing Newton coefficients for the expansion of the func from n_a coefficients to n_leja coefficients. Return a new value n_a=n_a+n_leja with the total number of Newton coefficients in the updated a.

Arguments

a: On input, a zero-based array of length n_a or greater, containing Newton coefficients. On output, array containing a total n_leja coefficients. The array a will be resized if necessary, and may have a length greater than n_leja on output

n_a: The number of Newton coefficients in a, on input. Elements of a beyond the first n_a elements will be overwritten.

leja: Array of normalized Leja points, containing at least n_leja elements.

func: Function for which to calcluate Newton coeffiecients

n_leja: The number of elements in leja to use for calculating new coefficients, and the total number of Newton coefficients on output

radius: Normalization radius for divided differences

QuantumPropagators.random_state — Method

Random normalized quantum state.

    Ψ = random_state(H)

returns a random normalized state compatible with the Hamiltonian H. This is intended to provide a starting vector for estimating the spectral radius of H via an Arnoldi method.

QuantumPropagators.ritzvals — Function

Calculate a vector for Ritz values converged to a given precision.

R = ritzvals(G, state, m_min, m_max=2*m_min; prec=1e-5, norm_min=1e-15)

calculates a complex vector R of at least m_min and at most m_max Ritz values.

Index

Krotov.KrotovResult

Krotov.PulseParametrization

QuantumControlBase.AbstractControlObjective

QuantumControlBase.ControlProblem

QuantumControlBase.Objective

QuantumControlBase.WeightedObjective

QuantumPropagators.ChebyWrk

QuantumPropagators.ExpPropWrk

QuantumPropagators.NewtonWrk

Base.adjoint

Krotov.NoParametrization

Krotov.SquareParametrization

Krotov.TanhParametrization

Krotov.TanhSqParametrization

Krotov.optimize_pulses

Krotov.print_table

QuantumControl.optimize

QuantumControlBase.F_re

QuantumControlBase.F_sm

QuantumControlBase.F_ss

QuantumControlBase.J_T_re

QuantumControlBase.J_T_sm

QuantumControlBase.J_T_ss

QuantumControlBase.blackman

QuantumControlBase.box

QuantumControlBase.chi_re!

QuantumControlBase.chi_sm!

QuantumControlBase.chi_ss!

QuantumControlBase.discretize

QuantumControlBase.discretize_on_midpoints

QuantumControlBase.f_tau

QuantumControlBase.flattop

QuantumControlBase.getcontrols

QuantumControlBase.initobjpropwrk

QuantumControlBase.liouvillian

QuantumControlBase.obj_genfunc

QuantumControlBase.setcontrolvals

QuantumControlBase.setcontrolvals!

QuantumPropagators.arnoldi!

QuantumPropagators.cheby!

QuantumPropagators.cheby_coeffs

QuantumPropagators.cheby_coeffs!

QuantumPropagators.diagonalize_hessenberg_matrix

QuantumPropagators.expprop!

QuantumPropagators.extend_arnoldi!

QuantumPropagators.extend_leja!

QuantumPropagators.extend_newton_coeffs!

QuantumPropagators.get_from_storage!

QuantumPropagators.init_storage

QuantumPropagators.initpropwrk

QuantumPropagators.map_observable

QuantumPropagators.map_observables

QuantumPropagators.newton!

QuantumPropagators.propagate

QuantumPropagators.propstep!

QuantumPropagators.random_state

QuantumPropagators.ritzvals

QuantumPropagators.specrange

QuantumPropagators.write_to_storage!