GRAPE Package

Index

GRAPE.GrapeResult

GRAPE.optimize_grape

GRAPE.print_table

GRAPE

Private

GRAPE.print_table — Function

Print optimization progress as a table.

This functions serves as the default info_hook for an optimization with GRAPE.

GRAPE.GrapeResult — Type

Result object returned by optimize_grape.

GRAPE.optimize_grape — Function

Optimize a control problem using GRAPE.

result = optimize_grape(problem)

optimizes the given control problem, returning a GrapeResult.

Note

It is recommended to call optimize with method=:GRAPE instead of calling optimize_grape directly.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives, τ=τ) that evaluates the final time functional from a list ϕ of forward-propagated states and problem.objectives.

gradient: A function gradient!(G, τ, ∇τ) that stores the gradient of J_T in G.

Optional problem keyword arguments

update_hook: Not immplemented

info_hook: A function that receives the same argumens as update_hook, in order to write information about the current iteration to the screen or to a file. The default info_hook prints a table with convergence information to the screen. Runs after update_hook. The info_hook function may return a tuple, which is stored in the list of records inside the GrapeResult object.

check_convergence: A function to check whether convergence has been reached. Receives a GrapeResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘. The convergence check is performed after any calls to update_hook and info_hook.

x_tol: Parameter for Optim.jl

f_tol: Parameter for Optim.jl

g_tol: Parameter for Optim.jl

show_trace: Parameter for Optim.jl

extended_trace: Parameter for Optim.jl

show_every: Parameter for Optim.jl

allow_f_increases: Parameter for Optim.jl

optimizer: An optional Optim.jl optimizer (Optim.AbstractOptimizer instance). If not given, an L-BFGS-B optimizer will be used.

prop_method/fw_prop_method/bw_prop_method: The propagation method to use for each objective, see below.

prop_method/fw_prop_method/grad_prop_method: The propagation method to use for the extended gradient vector for each objective, see below.

The propagation method for the forward propagation of each objective is determined by the first available item of the following:

a fw_prop_method keyword argument

a prop_method keyword argument

a property fw_prop_method of the objective

a property prop_method of the objective

the value :auto

The propagation method for the backword propagation is determined similarly, but with bw_prop_method instead of fw_prop_method. The propagation method for the forward propagation of the extended gradient vector for each objective is determined from grad_prop_method, fw_prop_method, prop_method in order of precedence.