QuantumPropagators Package

Index

QuantumPropagators.Cheby.ChebyWrk

QuantumPropagators.ExpProp.ExpPropWrk

QuantumPropagators.Newton.NewtonWrk

QuantumPropagators.Arnoldi.arnoldi!

QuantumPropagators.Arnoldi.diagonalize_hessenberg_matrix

QuantumPropagators.Arnoldi.extend_arnoldi!

QuantumPropagators.Cheby.cheby

QuantumPropagators.Cheby.cheby!

QuantumPropagators.Cheby.cheby_coeffs

QuantumPropagators.Cheby.cheby_coeffs!

QuantumPropagators.ExpProp.expprop!

QuantumPropagators.Newton.extend_leja!

QuantumPropagators.Newton.extend_newton_coeffs!

QuantumPropagators.Newton.newton!

QuantumPropagators.SpectralRange.random_state

QuantumPropagators.SpectralRange.ritzvals

QuantumPropagators.SpectralRange.specrange

QuantumPropagators.Storage.get_from_storage!

QuantumPropagators.Storage.init_storage

QuantumPropagators.Storage.map_observable

QuantumPropagators.Storage.map_observables

QuantumPropagators.Storage.write_to_storage!

QuantumPropagators.initpropwrk

QuantumPropagators.propagate

QuantumPropagators.propstep

QuantumPropagators.propstep!

QuantumPropagators

Public

QuantumPropagators.initpropwrk — Function

Initialize a workspace for propagation.

wrk = initpropwrk(state, tlist, method=:auto, generator...; kwargs...)

The resulting wrk can be passed to propagate or propstep!.

Arguments

state: An exemplary state for the propagation (e.g., the initial state)

tlist: The time grid over which propagate will be called. Must include at least to points in order to determine the propagation time step to prepare. If the propagation will be over a tlist with a variable dt, the full tlist must be passed here.

generator: An exemplary (non-time-dependent) dynamical generator. For full generality (if method=:cheby), the given generator should have a spectral range sufficiently large to encompass the entire propagation. If given multiple times, a spectral envelope enclosing all the generators will be determined automatically. In this case, you should pass the generators with the extremal values of all the controls.

method: The propagation method to use. The default value of :auto attempts to choose the best method available, based on the properties of the given state, tlist, and generator. Alternative values are :cheby and :newton, and :expprop.

specrad_method: for method=:cheby, method to use for estimating the spectral radius, see QuantumPropagators.SpectralRange.specrange. Defaults to :auto.

tolerance: for method=:cheby, a tolerance factor for the estimated spectral radius. That is, Chebychev coefficients will be calculated for a spectral radius increased by the tolerance factor compared to the specral radius estimated for the generator.

All other kwargs are filtered and passed to the contructor for returned workspace, e.g. limit for method=:cheby or m_max for method=:newton. For method=:cheby, they are also passed to specrange.

QuantumPropagators.propagate — Function

Propagate a state over an entire time grid.

state_out = propagate(
    state, genfunc, tlist; method=:auto,
    backwards=false; storage=nothing, observables=(<store state>, ),
    hook=nothing, showprogress=false, control_parameters=nothing,
    in_place=true, kwargs...)

propagates state over the time grid in tlist, using piecewise-constant dynamical generators (Hamiltonians or Liouvillians) determined by genfunc, and returns the resulting propagated state. The propagation is performed by calling propstep! for every interval in tlist, or propstep if in_place=false.

For the i'th time interval, genfunc(tlist, i) must return the generator for that time interval. Generally, when approximating a time-continuous dynamical generator as piecewise-constant on the time grid, it should be evaluated at the midpoint of the interval. A possible exception is the first and last interval, which may be better evaluated at tlist[1] and tlist[end] to ensure exact boundary conditions like control fields that are exactly zero.

In addition to the two positional parameters indicating the time interval, genfunc will also receive the state (the input state for the propagation step), backwards, storage, observables, control_parameters, and init as keyword arguments.

The control_parameters are an optional array of floats with parameters for genfunc. This is required when propagate is used in the context of automatic differentiation (AD). E.g., the Zygote framework can automatically calculate gradients of a function control_parameters -> J_T, where J_T might be be a function of the overlap between a propagated state (returned by propagate) and a target state. Thus, the control_parameters must be explicit in propagate. Outside of an AD context, control_parameters are not generally required: they can be implicit in genfunc.

The remaining keyword arguments may be used for unusual equations of motion beyond the standard Schrödinger or Liouville-von-Neumann equation, e.g. state would enter the genfunc for a Gross–Pitaevskii equation. For standard equations of motion that do not use the additional parameters, it is best to capture the keyword arguments to genfunc with a definition like

genfunc(tlist, i; kwargs...) = ...

For valid propagation methods, see initpropwrk.

In general, there is no requirement that tlist has a constant time step, although some propagation methods (most notably cheby!) only support a uniform time grid.

If storage is given as an Array, it will be filled with data determined by the observables. The default "observable" results in the propagated states at every point in time being stored. The storage array should be created with init_storage. See its documentation for details.

The storage parameter may also be given as true, and a new storage array will be created internally with init_storage and returned instead of the propagated state:

data = propagate(
    state, genfunc, tlist; method=:auto
    backwards=false; storage=true, observables=observables,
    hook=nothing, showprogress=false, kwargs...)

If backwards is true, the input state is assumed to be at time tlist[end], and the propagation progresses backwards in time (with a negative time step dt). If storage is given, it will be filled back-to-front during the backwards propagation.

If hook is given as a callable, it will be called after each propagation step, as hook(state, generator, tlist, i, wrk, observables) where i is the index of the time interval on tlist covered by the propagation step (0 for the initial state, respectives lastindex(tlist) for the backward propagation). The hook is called before calculating any observables. Example usage includes writing data to file, or modifying state, e.g., removing amplitude from the lowest and highest level to mitigate "truncation error".

If showprogress is given as true, a progress bar will be shown for long-running propagationn. In order to customize the progress bar, showprogress may also be a function that receives length(tlist) and returns a ProgressMeter.Progress instance.

If in_place=false is given, the propagation avoids in-place operations by using propstep instead of propstep!. This is often required in the context of automatic differentiation (AD), e.g., with Zygote. That is, use in_place=false if propagate is called inside a function to be passed to Zygote.gradient, Zygote.pullback, or a similar function. In and AD context, storage and showprogress should not be used.

The propagate routine returns the propagated state at tlist[end], respectively tlist[1] if backwards=true, or a storage array with the stored states / observable data if storage=true.

QuantumPropagators.propstep — Function

Perform a single propagation step and return the propagated state.

state_out = propstep(state, generator, dt, wrk; kwargs...)

The propagation method is determined by wrk, see initpropwrk.

Unlike propstep!, this method does not act in place, which generally makes it more suitable for automatic differentiation. However, there may be a performance penalty associated with the additional memory allocations.

The kwargs are forwarded to the underlying method.

QuantumPropagators.propstep! — Function

Perform a single propagation step in-place.

propstep!(state, generator, dt, wrk; kwargs...)

The propagation method is determined by wrk, see initpropwrk.

Generally, an in-place propagation will not be suitable for in the context of automatic differentiation. See propagate for a method that does not act in-place.

The kwargs are forwarded to the underlying method.

QuantumPropagators.Arnoldi

Private

QuantumPropagators.Arnoldi.diagonalize_hessenberg_matrix — Function

diagonalize_hessenberg_matrix(Hess, m; accumulate=false)

Diagonalize the m × m top left submatrix of the given Hessenberg matrix.

If accumulate is true, return the concatenated eigenvalues for Hess[1:1,1:1] to Hess[1:m,1:m], that is, all sumatrices of size 1 through m.

QuantumPropagators.Arnoldi.extend_arnoldi! — Function

Extend dimension of Hessenberg matrix by one.

extend_arnoldi!(Hess, q, m, H, dt; norm_min=1e-15)

extends the entries in Hess from size (m-1)×(m-1) to size m×m, and the list q of Arnoldi vectors from m to (m+1). It is assumed that the input Hess was created by a call to arnoldi! with extended=false or a previous call to extend_arnoldi!. Note that Hess itself is not resized, so it must be allocated to size m×m or greater on input.

QuantumPropagators.Arnoldi.arnoldi! — Function

m = arnoldi!(Hess, q, m, Ψ, H, dt=1.0; extended=true, norm_min=1e-15)

Calculate the Hessenberg matrix and Arnoldi vectors of H dt, from Ψ.

For a given order m, the m×m Hessemberg matrix is calculated and stored in in the pre-allocated Hess. Further an array of m normalized Arnoldi vectors is stored in in the pre-allocated q, plus one additional unnormalized Arnoldi vector. The unnormalized m+1st vector could be used to easily extend a given m×m Hessenberg matrix to a (m+1)×(m+1) matrix.

If the extended Hessenberg matrix is requested (extended=true, default), the m+1st Arnoldi vector is also normalized, and it's norm will be stored in m+1, m entry of the (extended) Hessenberg matrix, which is an (m+1)×(m+1) matrix.

Return the size m of the calculated Hessenberg matrix. This will usually be the input m, except when the Krylov dimension of H starting from Ψ is less then m. E.g., if Ψ is an eigenstate of H, the returned m will be 1.

See http://en.wikipedia.org/wiki/Arnoldi_iteration for a description of the algorithm.

Arguments

Hess::Matrix{ComplexF64}: Pre-allocated storage for the Hessemberg matrix. Can be uninitialized on input. The matrix must be at least of size m×m, or (m+1)×(m+1) if extended=true. On output, the m×m sub-matrix of Hess (with the returned output m) will contain the Hessenberg matrix, and all other elements of Hess be be set to zero.

q: Pre-allocated array of states similar to Ψ, as storage for the calculated Arnoldi vectors. These may be un-initialized on input. Must be at least of length m+1

m: The requested dimensions of the output Hessenberg matrix.

Ψ: The starting vector for the Arnoldi procedure. This can be of any type, as long as Φ = H * Ψ results in a vector similar to Ψ, there is an inner products of Φ and Ψ (Ψ⋅Φ is defined), and norm(Ψ) is defined.

H: The operator (up to dt) for which to calculate the Arnoldi procedure. Can be of any type, as long as H * Ψ is defined.

dt: The implicit time step; the total operator for which to calculate the Arnoldi procedure is H * dt

extended: If true (default), calculate the extended Hessenberg matrix, and normalized the final Arnoldi vector

norm_min: the minimum value of the norm of Ψ at which Ψ should be considered the zero vector

QuantumPropagators.Cheby

Public

QuantumPropagators.Cheby.ChebyWrk — Type

Workspace for the Chebychev propagation routine.

ChebyWrk(Ψ, Δ, E_min, dt; limit=1e-12)

initializes the workspace for the propagation of a state similar to Ψ under a Hamiltonian with eigenvalues between E_min and E_min + Δ, and a time step dt. Chebychev coefficients smaller than the given limit are discarded.

QuantumPropagators.Cheby.cheby — Function

Evaluate Ψ = exp(i- H dt) Ψ.

Ψ_out = cheby(Ψ, H, dt, wrk; E_min=nothing, check_normalization=false)

acts like cheby! but does not modify Ψ in-place.

QuantumPropagators.Cheby.cheby! — Function

Evaluate Ψ = exp(-i H dt) Ψ in-place.

cheby!(Ψ, H, dt, wrk; E_min=nothing, check_normalization=false)

Arguments

Ψ: on input, initial vector. Will be overwritten with result.

H: Hermitian operator

dt: time step

wrk: internal workspace

E_min: minimum eigenvalue of H, to be used instead of the E_min from the initialization of wrk. The same wrk may be used for different values E_min, as long as the spectra radius Δ and the time step dt are the same as those used for the initialization of wrk.

check_normalizataion: perform checks that the H does not exceed the spectral radius for which the the workspace was initialized.

The routine will not allocate any internal storage. This implementation requires copyto! lmul!, and axpy! to be implemented for Ψ, and the three-argument mul! for Ψ and H.

QuantumPropagators.Cheby.cheby_coeffs — Function

Calculate Chebychev coefficients.

a::Vector{Float64} = cheby_coeffs(Δ, dt; limit=1e-12)

return an array of coefficiencts larger than limit.

Arguments

Δ: the spectral radius of the underlying operator

dt: the time step

See also cheby_coeffs! for an in-place version.

QuantumPropagators.Cheby.cheby_coeffs! — Function

Calculate Chebychev coefficients in-place.

n::Int = cheby_coeffs!(coeffs, Δ, dt, limit=1e-12)

overwrites the first n values in coeffs with new coefficients larger than limit for the given new spectral radius Δ and time step dt. The coeffs array will be resized if necessary, and may have a length > n on exit.

See also cheby_coeffs for an non-in-place version.

QuantumPropagators.ExpProp

Public

QuantumPropagators.ExpProp.ExpPropWrk — Type

ExpPropWrk(v0)

Workspace for propagation via direct matrix exponentiation.

Initializes the workspace for the propagation of a vector v0

QuantumPropagators.ExpProp.expprop! — Function

expprop!(Ψ, H, dt, wrk; func=(H_dt -> exp(-1im * H_dt)))

Evaluate Ψ = func(H*dt) Ψ by directly evaluating U = func(H*dt), i.e. by matrix exponentiation for the default func, and then multiplying U and Ψ in-place with mul!.

The workspace wrk must be initialized with ExpPropWrk to provide storage for a temporary state.

QuantumPropagators.Newton

Public

QuantumPropagators.Newton.NewtonWrk — Type

NewtonWrk(v0, m_max=10)

Workspace for the Newton-with-restarted-Arnoldi propagation routine.

Initializes the workspace for the propagation of a vector v0, using a maximum Krylov dimension of m_max in each restart iteration. Note that m_max should be smaller than the length of v0.

QuantumPropagators.Newton.newton! — Function

newton!(Ψ, H, dt, wrk; func=(z -> exp(-1im*z)), norm_min=1e-14, relerr=1e-12,
        max_restarts=50)

Evaluate Ψ = func(H*dt) Ψ using a Newton-with-restarted-Arnoldi scheme.

Arguments

Ψ: The state to propagate, will be overwritten in-place with the propagated state

H: Operator acting on Ψ. Together with dt, this is the argument to func

dt: Implicit time step. Together with H, this is the argument to func

wkr: Work array, initialized with NewtonWrk

func: The function to apply to H dt, taking a single (scalar) complex-valued argument z in place of H dt. The default func is to evaluate the time evoluation operator for the Schrödinger equation

norm_min: the minium norm at which to consider a state similar to Ψ as zero

relerr: The relative error defining the convergence condition for the restart iteration. Propagation stops when the norm of the accumulated Ψ is stable up to the given relative error

max_restart: The maximum number of restart iterations. Exceeding max_restart will throw an AssertionError.

Private

QuantumPropagators.Newton.extend_newton_coeffs! — Function

extend_newton_coeffs!(a, n_a, leja, func, n_leja, radius)

Extend the array a of existing Newton coefficients for the expansion of the func from n_a coefficients to n_leja coefficients. Return a new value n_a=n_a+n_leja with the total number of Newton coefficients in the updated a.

Arguments

a: On input, a zero-based array of length n_a or greater, containing Newton coefficients. On output, array containing a total n_leja coefficients. The array a will be resized if necessary, and may have a length greater than n_leja on output

n_a: The number of Newton coefficients in a, on input. Elements of a beyond the first n_a elements will be overwritten.

leja: Array of normalized Leja points, containing at least n_leja elements.

func: Function for which to calcluate Newton coeffiecients

n_leja: The number of elements in leja to use for calculating new coefficients, and the total number of Newton coefficients on output

radius: Normalization radius for divided differences

QuantumPropagators.Newton.extend_leja! — Function

extend_leja!(leja, n, newpoints, n_use)

Given an array of n (ordered) Leja points, extract n_use points from newpoints, and append them to the existing Leja points. The array leja should be sufficiently large to hold the new Leja points, which are appended after index n_old. It will be re-allocated if necessary and may have a size of up to 2*(n+n_use).

Arguments

leja: Array of leja values. Must contain the "old" leja values to be kept in leja(0:n-1). On output, n_use new leja points will be in leja(n+:n+n_use-1), for the original value of n. The leja array must use zero-based indexing.

n: On input, number of "old" leja points in leja. On output, total number of leja points (i.e. n=n+n_use)

newpoints: On input, candidate points for new leja points. The n_use best values will be chosen and added to leja. On output, the values of new_points are undefined.

n_use: Number of points that should be added to leja

QuantumPropagators.SpectralRange

Public

QuantumPropagators.SpectralRange.specrange — Function

Calculate the spectral range of a Hamiltonian H on the real axis.

E_min, E_max = specrange(H; method=:auto, kwargs...)

calculates the approximate lowest and highest eigenvalues of H. Any imaginary part in the eigenvalues is ignored: the routine is intended for (although not strictly limited to) a Hermitian H.

This delegates to

specrange(H, method; kwargs...)

for the different methods.

The default method=:auto chooses the best method for the given H. This is :diag for small matrices, and :arnoldi otherwise. Keyword arguments not relevant to the underlying implementation will be ignored.

E_min, E_max = specrange(H, :arnoldi; state=random_state(H), m_min=20,
                         m_max=60, prec=1e-3, nom_min=1e-15, enlarge=true)

uses Arnoldi iteration with state as the starting vector. It approximates the eigenvalues of H with between m_min and m_max Ritz values, until the lowest and highest eigenvalue are stable to a releative precision of prec. The norm_min parameter is passed to the underlying arnoldi!.

If enlarge=true (default) the returned E_min and E_max will be enlarged via a heuristic to slightly over-estimate the spectral radius instead of under-estimating it.

E_min, E_max = specrange(H, :diag)

uses exact diagonization via the standard eigenvals function to obtain the smallest and largest eigenvalue. This should only be used for relatively small matrices.

Private

QuantumPropagators.SpectralRange.ritzvals — Function

Calculate a vector for Ritz values converged to a given precision.

R = ritzvals(G, state, m_min, m_max=2*m_min; prec=1e-5, norm_min=1e-15)

calculates a complex vector R of at least m_min and at most m_max Ritz values.

QuantumPropagators.SpectralRange.random_state — Function

Random normalized quantum state.

    Ψ = random_state(H)

returns a random normalized state compatible with the Hamiltonian H. This is intended to provide a starting vector for estimating the spectral radius of H via an Arnoldi method.

QuantumPropagators.Storage

Public

QuantumPropagators.Storage.get_from_storage! — Function

Obtain data from storage.

get_from_storage!(state, storage, i)

extracts data from the storage for the i'th time slot. Invese of write_to_storage!

QuantumPropagators.Storage.init_storage — Function

Create a storage array for propagation.

storage = init_storage(state, tlist)

creates a storage array suitable for storing a state for each point in tlist.

storage = init_storage(state, tlist, observables))

creates a storage array suitable for the data generated by the observables applied to state, see map_observables, for each point in tlist.

storage = init_storage(data, nt))

creates a storage arrays suitable for storing data nt times, where nt=length(tlist). By default, this will be a vector of typeof(data) and length nt, or a n × nt Matrix with the same eltype as data if data is a Vector of length n.

QuantumPropagators.Storage.map_observable — Function

Apply a single observable to state.

data = map_observable(observable, state)

By default, observable is assumed to be callable, and the above is equivalent to data = observable(state).

If observable is a matrix and state is a vector evaluate the expectation value of the observable as dot(state, observable, state).

QuantumPropagators.Storage.map_observables — Function

Obtain "observable" data from state.

data = map_observables(observables, state)

calculates the data for a tuple of observables applied to state. For a single observable (tuple of length 1), simply return the result of map_observable.

For multiple observables, return the tuple resulting from applying map_observable for each observable. If the tuple is "uniform" (all elements are of the same type, e.g. if each observable calculates the expectation value of a Hermitian operator), it is converted to a Vector. This allows for compact storage in a storage array, see init_storage.

QuantumPropagators.Storage.write_to_storage! — Function

Place data into storage for time slot i.

write_to_storage!(storage, i, state, observables)

For a storage array created by init_storage, store the data obtains from map_observables into the storage for time slot i. This delegates to the more general

write_to_storage!(storage, i, data)

Conceptually, this corresponds roughly to storage[i] = data, but storage may have its own idea on how to store data for a specific time slot. For example, with the default init_storage Vector data will be stored in a matrix, and write_to_storage! will in this case write data to the i'th column of the matrix.

For a given type of storage and data, it is the developer's responsibility that init_storage and write_to_storage! are compatible.