GRAPE Package

Index

$\gdef\tgt{\text{tgt}}$ $\gdef\tr{\operatorname{tr}}$ $\gdef\Re{\operatorname{Re}}$ $\gdef\Im{\operatorname{Im}}$

GRAPE.GrapeResult

GRAPE.optimize_grape

GRAPE.print_table

GRAPE

Private

GrapeResult

print_table

optimize_grape

GRAPE.GrapeResult — Type

Result object returned by optimize_grape.

GRAPE.print_table — Function

Print optimization progress as a table.

This functions serves as the default info_hook for an optimization with GRAPE.

GRAPE.optimize_grape — Function

Optimize a control problem using GRAPE.

result = optimize_grape(problem)

optimizes the given control problem, returning a GrapeResult.

Note

It is recommended to call optimize with method=:GRAPE instead of calling optimize_grape directly.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives; τ=τ) that evaluates the final time functional from a vector ϕ of forward-propagated states and problem.objectives. For all objectives that define a target_state, the element τₖ of the vector τ will contain the overlap of the state ϕₖ with the target_state of the k'th objective, or NaN otherwise.

Optional problem keyword arguments

gradient_via: A flag indicating how the gradient of J_T should be calculated. One of :tau, :chi. If all objectives in problem define a target_function, the default is gradient_via=:tau. This understands the functional J_T as a function of the complex overlaps between the propagated states and the target states, and evaluates the gradient via the chain rule. For functionals that cannot be expressed in terms of the overlap of propagated and target states, and/or if not all objectives define a target state, gradient_via=:chi because the default. In this case, the functional J_T is understood as a function of the forward-propagated states directly, and the full gradient is again calculated by the chain rule. See make_gradient for details.

gradient: A function to evaluate the gradient of J_T. By default, it is constructed via make_gradient. If given manually, it must meet the interface described by make_gradient. Most importantly, it must be consistent with the chosen gradient_via.

chi: If gradient_via=:chi, a function that constructs the χ-states for the backward propagation, see make_gradient for details. By default, it is constructed via make_chi. If given manually, it must meet the same interface described in make_chi.

force_zygote=false: Whether to force the use of automatic differentiation in make_gradient and make_chi. This disables analytic gradients. The only reason to do this is for testing/benchmarking analytic vs automatic gradients.

update_hook: Not implemented

info_hook: A function that receives the same arguments as update_hook, in order to write information about the current iteration to the screen or to a file. The default info_hook prints a table with convergence information to the screen. Runs after update_hook. The info_hook function may return a tuple, which is stored in the list of records inside the GrapeResult object.

check_convergence: A function to check whether convergence has been reached. Receives a GrapeResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘. The convergence check is performed after any calls to update_hook and info_hook.

x_tol: Parameter for Optim.jl

f_tol: Parameter for Optim.jl

g_tol: Parameter for Optim.jl

show_trace: Parameter for Optim.jl

extended_trace: Parameter for Optim.jl

show_every: Parameter for Optim.jl

allow_f_increases: Parameter for Optim.jl

optimizer: An optional Optim.jl optimizer (Optim.AbstractOptimizer instance). If not given, an L-BFGS-B optimizer will be used.

prop_method/fw_prop_method/bw_prop_method: The propagation method to use for each objective, see below.

prop_method/fw_prop_method/grad_prop_method: The propagation method to use for the extended gradient vector for each objective, see below.

verbose=false: If true, print information during initialization

The propagation method for the forward propagation of each objective is determined by the first available item of the following:

a fw_prop_method keyword argument

a prop_method keyword argument

a property fw_prop_method of the objective

a property prop_method of the objective

the value :auto

The propagation method for the backward propagation is determined similarly, but with bw_prop_method instead of fw_prop_method. The propagation method for the forward propagation of the extended gradient vector for each objective is determined from grad_prop_method, fw_prop_method, prop_method in order of precedence.