QuantumControlBase Package

Index

QuantumControlBase.AbstractControlObjective

QuantumControlBase.ControlProblem

QuantumControlBase.GradGenerator

QuantumControlBase.GradVector

QuantumControlBase.Objective

QuantumControlBase.TimeDependentGradGenerator

QuantumControlBase.WeightedObjective

QuantumControlBase.Functionals.F_re

QuantumControlBase.Functionals.F_sm

QuantumControlBase.Functionals.F_ss

QuantumControlBase.Functionals.J_T_re

QuantumControlBase.Functionals.J_T_sm

QuantumControlBase.Functionals.J_T_ss

QuantumControlBase.Functionals.chi_re!

QuantumControlBase.Functionals.chi_sm!

QuantumControlBase.Functionals.chi_ss!

QuantumControlBase.Functionals.f_tau

QuantumControlBase.Functionals.grad_J_T_sm!

QuantumControlBase.Shapes.blackman

QuantumControlBase.Shapes.box

QuantumControlBase.Shapes.flattop

QuantumControlBase.TestUtils.random_complex_matrix

QuantumControlBase.TestUtils.random_complex_sparse_matrix

QuantumControlBase.TestUtils.random_hermitian_matrix

QuantumControlBase.TestUtils.random_hermitian_sparse_matrix

QuantumControlBase.TestUtils.random_real_matrix

QuantumControlBase.TestUtils.random_real_sparse_matrix

QuantumControlBase.TestUtils.random_state_vector

QuantumControlBase.chain_infohooks

QuantumControlBase.default_optimization_savename_kwargs

QuantumControlBase.discretize

QuantumControlBase.discretize_on_midpoints

QuantumControlBase.evalcontrols

QuantumControlBase.evalcontrols!

QuantumControlBase.get_control_parameters

QuantumControlBase.get_tlist_midpoints

QuantumControlBase.getcontrolderiv

QuantumControlBase.getcontrolderivs

QuantumControlBase.getcontrols

QuantumControlBase.initobjpropwrk

QuantumControlBase.liouvillian

QuantumControlBase.load_optimization

QuantumControlBase.objective_genfunc

QuantumControlBase.optimization_savename

QuantumControlBase.optimize

QuantumControlBase.propagate_objective

QuantumControlBase.resetgradvec!

QuantumControlBase.@optimize_or_load

QuantumControlBase.ConditionalThreads.@threadsif

QuantumControlBase

Public

QuantumControlBase.@optimize_or_load — Macro

Run optimize and store the result, or load the result if it exists.

result, file = @optimize_or_load(
    path="",
    problem;
    method=<method>,
    filename=nothing,
    suffix="jld2",
    prefix=DrWatson.default_prefix(config),
    tag=DrWatson.readenv("DRWATSON_TAG", true),
    gitpath=DrWatson.projectdir(),
    storepatch::Bool=DrWatson.readenv("DRWATSON_STOREPATCH", false),
    force=false,
    verbose=true,
    wsave_kwargs=Dict(),
    savename_kwargs=DEFAULT_OPTIMIZATION_SAVENAME_KWARGS,
    kwargs...
)

runs result = optimize(problem; method=<method>, kwargs...) and stores result in an automatically named file inside path. The automatic file name is determined by optimization_savename and can be overriden by passing an explicit filename. The full path to the output file (joinpath(path, filename)) is returned as file.

In addition to the result, the data in the output file may also contain some metadata, e.g. "gitcommit" containing the git commit hash of the project the produced the file, and "script" with the file name and line number where @optimize_or_load was called, see load_optimization.

If file already exists (and force=false), load the result from that file instead of running the optimization.

The @optimize_or_load macro is intended to integrate well with the DrWatson framework for scientific projects and utilizes several configuration options and utility functions from DrWatson, see below. Note that even though DrWatson is recomended, you are not required to use if for your projects in order to use @optimize_or_load or any other part of QuantumControl.

I/O Keywords

The following keyword arguments determine where the result is stored and in which format.

filename: A file name to override the automatic file name. The filename should not contain slashes: use path for the folder where filename should be created.

suffix, prefix, savename_kwargs: Parameters for optimization_savename, which determines the automatic file name

tag: Whether to record the current "gitcommit" as metadata alongside the optimization result, via DrWatson.tagsave. If not given explicitly, determine automatically from suffix or the extension of filename.

gitpath, storepatch: Passed to DrWatson.tagsave if tag is true.

force: If true, run and store the optimization regardless of whether file already exists.

verbose: If true, print info about the process, if file does not exist.

wsave_kwargs: Additional keyword arguments to pass to DrWatson.wsave, e.g., to enable compression

All other keyword arguments are passed directly to optimize.

Related Functions

optimization_savename: Function that determines the automatic filename

DrWatson.@produce_or_load: The lower-level backend implementing the functionality of @optimize_or_load.

load_optimization: Function to load a file produced by @optimize_or_load

QuantumControlBase.ControlProblem — Type

A full control problem with multiple objectives.

ControlProblem(
   objectives=<list of objectives>,
   pulse_options=<dict of controls to pulse options>,
   tlist=<time grid>,
   kwargs...
)

Note that the control problem can only be instantiated via keyword arguments.

The objectives are a list of AbstractControlObjective instances, each defining an initial state and a dynamical generator for the evolution of that state. Usually, the objective will also include a target state (see Objective) and possibly a weight (see WeightedObjective).

The pulse_options are a dictionary (IdDict) mapping controls that occur in the objectives to properties specific to the control method.

The tlist is the time grid on which the time evolution of the initial states of each objective should be propagated.

The remaining kwargs are keyword arguments that are passed directly to the optimal control method. These typically include e.g. the optimization functional.

The control problem is solved by finding a set of controls that simultaneously fulfill all objectives.

QuantumControlBase.Objective — Type

Standard optimization objective.

Objective(;
    initial_state=<initial_state>,
    generator=<genenerator>,
    target_state=<target_state>
)

describes an optimization objective where the time evoluation of the given initial_state under the given generator aims towards target_state. The generator here is e.g. a time-dependent Hamiltonian or Liouvillian.

The most common control problems in quantum control, e.g. state-to-state transitions or quantum gate implementations can be expressed by simultaneously fulfilling multiple objectives of this type.

Note that the objective can only be instantiated via keyword arguments.

QuantumControlBase.WeightedObjective — Type

Standard optimization objective with a weight.

WeightedObjective(;
    initial_state=<initial_state>,
    generator=<genenerator>,
    target_state=<target_state>,
    weight=<weight>
)

initializes a control objective like Objective, but with an additional weight parameter (a float generally between 0 and 1) that weights the objective relative to other objectives that are part of the same control problem.

QuantumControlBase.chain_infohooks — Function

Combine multiple info_hook functions.

chain_infohooks(funcs...)

combines funcs into a single Function that can be passes as info_hook to ControlProblem or any optimize-function.

Each function in func must be a suitable info_hook by itself. This means that it should receive the optimization workspace object as its first positional parameter, then positional parameters specific to the optimization method, and then an arbitrary number of data parameters. It must return either nothing or a tuple of "info" objects (which will end up in the records field of the optimization result).

When chaining infohooks, the funcs will be called in series, and the "info" objects will be accumulated into a single result tuple. The combined results from previous funcs will be given to the subsequent funcs as data parameters. This allows for the infohooks in the chain to communicate.

The chain will return the final combined result tuple, or nothing if all funcs return nothing.

QuantumControlBase.default_optimization_savename_kwargs — Function

Set the default savename_kwargs for optimization_savename.

savename_kwargs = default_optimization_savename_kwargs(;kwargs...)

sets entries in the DEFAULT_OPTIMIZATION_SAVENAME_KWARGS used in optimization_savename and thus determines the automatic name used to store optimization results.

Use

default_optimization_savename_kwargs(reset=true)

to clear the settings from any previous call to default_optimization_savename_kwargs.

The following keyword arguments are supported, cf. DrWatson.savename:

accesses - List of strings indicating which fields (keys in kwargs of ControlProblem or kwargs of optimize/@optimize_or_load) can be included in the output filename. By default, all fields with values matching allowedtypes (excluding anonymous functions) are used.

allowedtypes - List of types of values eligible to be included in the filename. Defaults to [Real, String, Symbol, TimeType, Function]

connector - String used to separate key-value pairs in the output filename. Defaults to "_".

digits - Used in round when formatting numbers, if no custom val_to_string.

equals - String used between keys and values. Defaults to "=".

ignores - List of strings indicating which fields should be ignored.

sigdigits - Used in round when formatting numbers, if no custom val_to_string.

sort - Whether to sort the fields alphabetically (default). If false, the resulting file name my not be stable.

val_to_string - Function to convert values to string.

QuantumControlBase.discretize — Function

Evaluate control at every point of tlist.

values = discretize(control, tlist; via_midpoints=true)

discretizes the given control to a Vector of values defined on the points of tlist.

If control is a function, it will will first be evaluated at the midpoint of tlist, see discretize_on_midpoints, and then the values on the midpoints are converted to values on tlist. This discretization is more stable than directly evaluationg the control function at the values of tlist, and ensures that repeated round-trips between discretize and discretize_on_midpoints can be done safely, see the note in the documentation of discretize_on_midpoints.

The latter can still be achieved by passing via_midpoints=false. While such a direct discretization is suitable e.g. for plotting, but it is unsuitable for round-trips between discretize and discretize_on_midpoints (constant controls on tlist may result in a zig-zag on the intervals of tlist).

If control is a vector, it will be returned un-modified if it is of the same length as tlist. Otherwise, control must have one less value than tlist, and is assumed to be defined on the midpoins of tlist. In that case, discretize acts as the inverse of discretize_on_midpoints. See discretize_on_midpoints for how control values on tlist and control values on the intervals of tlist are related.

QuantumControlBase.discretize_on_midpoints — Function

Evaluate control at the midpoints of tlist.

values = discretize_on_midpoints(control, tlist)

discretizes the given control to a Vector of values on the midpoints of tlist. Hence, the resulting values will contain one less value than tlist.

If control is a vector of values defined on tlist (i.e., of the same length as tlist), it will be converted to a vector of values on the intervals of tlist. The value for the first and last "midpoint" will remain the original values at the beginning and end of tlist, in order to ensure exact bounary conditions. For all other midpoints, the value for that midpoint will be calculated by "un-averaging".

For example, for a control and tlist of length 5, consider the following diagram:

tlist index:       1   2   3   4   5
tlist:             ⋅   ⋅   ⋅   ⋅   ⋅   input values cᵢ (i ∈ 1..5)
                   |̂/ ̄ ̄ ̂\ / ̂\ / ̂ ̄ ̄\|̂
midpoints:         x     x   x     x   output values pᵢ (i ∈ 1..4)
midpoints index:   1     2   3     4

We will have $p₁=c₁$ for the first value, $p₄=c₅$ for the last value. For all other points, the control values $cᵢ = \frac{p_{i-1} + p_{i}}{2}$ are the average of the values on the midpoints. This implies the "un-averaging" for the midpoint values $pᵢ = 2 c_{i} - p_{i-1}$.

Note

An arbitrary input control array may not be compatible with the above averaging formula. In this case, the conversion will be "lossy" (discretize will not recover the original control array; the difference should be considered a "discretization error"). However, any further round-trip conversions between points and intervals are bijective and preserve the boundary conditions. In this case, the discretize_on_midpoints and discretize methods are each other's inverse. This also implies that for an optimal control procedure, it is safe to modify midpoint values. Modifying the the values on the time grid directly on the other hand may accumulate discretization errors.

If control is a vector of one less length than tlist, it will be returned unchanged, under the assumption that the input is already properly discretized.

If control is a function, the function will be directly evaluated at the midpoints marked as x in the above diagram..

QuantumControlBase.get_control_parameters — Function

Collect all control parameters from the given objectives.

get_control_parameters(objectives...; kwargs...)

returns a vector of control parameters extracted from the controls in objectives. This first extracts the controls with [getcontrols] and then the "control parameters" from each control. What are the "control parameters" are depends on the type of the control. For controls that are function ϵ(t), the control parameters are the values of the control function on intervals of the time grid. The time grid in the case must be passed as a keyword argument tlist.

The control parameters for multiple controls will be concatenated in the returned vector.

get_control_parameters(func; tlist, on_midpoints=true)

returns the control parameters of a control function by discretizing to the time grid in tlist. By default, the discretization is on intervals (midpoints) of the time grid. With on_midpoints=false, the discretization will be on the actual points of the time grid.

QuantumControlBase.get_tlist_midpoints — Function

Shift time grid values the interval midpoints

tlist_midpoints = get_tlist_midpoints(tlist)

takes a vector tlist of length $n$ and returns a vector of length $n-1$ containing the midpoint values of each interval. The intervals in tlist are not required to be uniform.

QuantumControlBase.getcontrols — Function

Extract a Tuple of controls.

controls = getcontrols(generator)

extracts the controls from a single dynamical generator.

controls = getcontrols(objectives)

extracts the controls from a list of objectives (i.e., from each objective's generator)

In either case, controls that occur multiple times, either in a single generator, or throughout the different objectives, will occur only once in the result.

By default, assumes that any generator is a nested Tuple, e.g. (H0, (H1, ϵ1), (H2, ϵ2), ...) and extracts (ϵ1, ϵ2)

Each control must be a valid argument for discretize.

QuantumControlBase.liouvillian — Function

Construct a Liouvillian super-operator.

ℒ = liouvillian(Ĥ, c_ops=(); convention=:LvN)

calculates the sparse Liouvillian super-operator ℒ from the Hamiltonian Ĥ and a list c_ops of Lindblad operators.

With convention=:LvN, applying the resulting ℒ to a vectorized density matrix ρ⃗ calculates $\frac{d}{dt} \vec{\rho}(t) = ℒ \vec{\rho}(t)$ equivalent to the Liouville-von-Neumann equation for the density matrix $ρ̂$,

\[\frac{d}{dt} ρ̂(t) = -i [Ĥ, ρ̂(t)] + \sum_k\left( Â_k ρ̂ Â_k^\dagger - \frac{1}{2} A_k^\dagger Â_k ρ̂ - \frac{1}{2} ρ̂ Â_k^\dagger Â_k \right)\,,\]

where the Lindblad operators $Â_k$ are the elements of c_ops.

The Hamiltonian $Ĥ$ may be time-dependent, using a nested-tuple format by default, e.g., (Ĥ₀, (H₁, ϵ₁), (H₂, ϵ₂)), where ϵ₁ and ϵ₂ are functions of time. In this case, the resulting ℒ will also be in nested tuple format, ℒ = (ℒ₀, (ℒ₁, ϵ₁), (ℒ₂, ϵ₂)), where the initial element contains the superoperator ℒ₀ for the static component of the Liouvillian, i.e., the commutator with the drift Hamiltonian Ĥ₀, plus the dissipator (sum over $k$), as a sparse matrix. Time-dependent Lindblad operators are not supported. The remaining elements are tuples (ℒ₁, ϵ₁) and (ℒ₂, ϵ₂) corresponding to the commutators with the two control Hamiltonians, where ℒ₁ and ℒ₂ again are sparse matrices.

If $Ĥ$ is not time-dependent, the resulting ℒ will be a single-element tuple containing the Liouvillian as a sparse matrix, ℒ = (ℒ₀, ).

With convention=:TDSE, the Liouvillian will be constructed for the equation of motion $-i \hbar \frac{d}{dt} \vec{\rho}(t) = ℒ \vec{\rho}(t)$ to match exactly the form of the time-dependent Schrödinger equation. While this notation is not standard in the literature of open quantum systems, it has the benefit that the resulting ℒ can be used in a numerical propagator for a (non-Hermitian) Schrödinger equation without any change. Thus, for numerical applications, convention=:TDSE is generally preferred. The returned ℒ between the two conventions differs only by a factor of $i$, since we generally assume $\hbar=1$.

The convention keyword argument is mandatory, to force a conscious choice.

See Goerz et. al. "Optimal control theory for a unitary operation under dissipative evolution", arXiv 1312.0111v2, Appendix B.2 for the explicit construction of the Liouvillian superoperator as a sparse matrix.

QuantumControlBase.load_optimization — Function

Load a previously stored optimization.

result = load_optimization(filename; verbose=true, kwargs...)

recovers a result previously stored by @optimize_or_load.

result, metadata = load_optimization(filename; return_metadata=true, kwargs...)

also obtains a metadata dict containing e.g., "gitcommit" or "script" depending on the options to @optimize_or_load.

Calling load_optimization with verbose=true (default) will show the metadata after loading the file.

QuantumControlBase.optimization_savename — Function

Determine an automatic filename for storing an optimization result.

file = optimization_savename(
    path="",
    problem;
    method=<method>,
    suffix="jld2",
    prefix="",
    savename_kwargs=DEFAULT_OPTIMIZATION_SAVENAME_KWARGS,
    kwargs...,
)

finds an appropriate automatic filename for the result of optimize(problem; method=<method>, kwargs...).

By default, the file has the structure <path>/<prefix>_<key1>=<value1>_..._<keyN>=<valueN>_method=<method>.jld2 where the key-value pairs are a subset of the keyword arguments used to instantiate problem, respectively the keyword arguments in kwargs. The prefix is best used as a "name" for the optimization problem to ensure a unique file name.

Which key-value pairs that are taken into account and the way they are formatted can be customized via savename_kwargs. See default_optimization_savename_kwargs for the supported options.

QuantumControlBase.optimize — Function

Optimize a quantum control problem.

opt_result = optimize(problem; method=<method>, kwargs...)

optimizes towards a solution of given problem with the given optimization method. Any keyword argument temporarily overrides the corresponding keyword argument in problem.

opt_result = optimize(problem; method=:krotov, kwargs...)

optimizes problem using Krotov's method, see Krotov.optimize_krotov.

opt_result = optimize(problem; method=:GRAPE, kwargs...)

optimizes problem using GRadident Ascent Pulse Engineering (GRAPE), see GRAPE.optimize_grape.

QuantumControlBase.propagate_objective — Function

Propagate with the dynamical generator of a control objective.

propagate_objective(obj, tlist; method=:auto, initial_state=obj.initial_state,
                    controls_map=IdDict(), kwargs...)

propagates initial_state under the dynamics described by obj.generator.

The optional dict control_map may be given to replace the controls in obj.generator (as obtained by getcontrols) with custom functions or vectors, e.g. with the controls resulting from optimization.

If obj has a property/field prop_method or fw_prop_method, its value will be used as the default for method instead of :auto. An explicit keyword argument for method always overrides the default.

All other kwargs are forwarded to the underlying QuantumPropagators.propagate method for obj.initial_state.

Private

QuantumControlBase.AbstractControlObjective — Type

Base class for a single optimization objective.

All objectives must have a field initial_state and a field generator, at minimum.

QuantumControlBase.resetgradvec! — Function

Reset the given gradient vector for a new gradient evaluation.

resetgradvec!(Ψ̃::GradVector)

zeroes out Ψ̃.grad_states but leaves Ψ̃.state unaffected.

resetgradvec!(Ψ̃::GradVector, Ψ)

additionally sets Ψ̃.state to Ψ.

QuantumControlBase.objective_genfunc — Function

Construct a genfunc suitable for propagating an objective.

genfunc = objective_genfunc(obj, tlist; controls_map=IdDict())

can be passed to QuantumPropagators.propagate to propagate under the dynamical generator in obj.

If given, control_map defines replacements for the controls in obj.generator. This allows, e.g., to replace the controls with those resulting from an optimization.

QuantumControlBase.getcontrolderiv — Function

Get the derivative of the generator $G$ w.r.t. the control $ϵ(t)$.

μ  = getcontrolderiv(generator, control)

returns nothing if the generator (Hamiltonian or Liouvillian) does not depend on control, or a function μ(v) that evaluates

\[μ(v) = \left.\frac{∂G}{∂ϵ(t)}\right\vert_{ϵ(t)=v}\]

otherwise. That is, a call μ(v) will return the static operator resulting from evaluating the derivative of the dynamical generator $G$ with respect to the control filed $ϵ(t)$ at a particular point in time where the control field takes the value $v$.

Note that for the common case of linear control terms, e.g., $Ĥ = Ĥ_0 + \sum_l ϵ_l(t) Ĥ_l$, the derivative $∂Ĥ/∂ϵ_l(t)$ is simply the control Hamiltonian $Ĥ_l$. Thus, the resulting function μ will simply return $Ĥ_l$, ignoring the argument v.

QuantumControlBase.TimeDependentGradGenerator — Type

Extended generator for the standard dynamic gradient.

G̃ = TimeDependentGradGenerator(G)

contains the original time-dependent generator G (a Hamiltonian or Liouvillian) in G̃.G, a vector of control derivatives $∂G/∂ϵₗ(t)$ in G̃.control_derivs, and the controls in G̃.controls.

For a generator $G = Ĥ(t) = Ĥ₀ + ϵ₁(t) Ĥ₁ + … + ϵₙ(t) Ĥₙ$, this extended generator encodes the block-matrix

\[G̃ = \begin{pmatrix} Ĥ(t) & 0 & \dots & 0 & Ĥ₁ \\ 0 & Ĥ(t) & \dots & 0 & Ĥ₂ \\ \vdots & & \ddots & & \vdots \\ 0 & 0 & \dots & Ĥ(t) & Ĥₙ \\ 0 & 0 & \dots & 0 & Ĥ(t) \end{pmatrix}\]

Note that the $∂G/∂ϵₗ(t)$ ($Ĥₗ$ in the above example) are functions, to account for the possibility of non-linear control terms, see getcontrolderiv.

QuantumControlBase.GradGenerator — Type

Static generator for the standard dynamic gradient.

G::GradGenerator = evalcontrols(G_of_t::TimeDependentGradGenerator, vals_dict)

is the result of plugging in specific values for all controls in a TimeDependentGradGenerator. See evalcontrols and evalcontrols!.

The resulting object can be multiplied directly with a GradVector, e.g., in the process of evaluating a piecewise-constant time propagation.

QuantumControlBase.evalcontrols! — Function

In-place version of evalcontrols.

evalcontrols!(G, generator, vals_dict)

acts as evalcontrols, but modifies G in-place.

QuantumControlBase.evalcontrols — Function

Replace the controls in generator with static values.

G = evalcontrols(generator, vals_dict)

replaces the time-dependent controls in generator with the values in vals_dict and returns the static operator G.

The vals_dict is a dictionary (IdDict) mapping controls as returned by getcontrols(generator) to values.

QuantumControlBase.GradVector — Type

Extended state-vector for the dynamic gradient.

Ψ̃ = GradVector(Ψ, num_controls)

for an initial state Ψ and num_controls control fields.

The GradVector conceptually corresponds to a direct-sum (block) column-vector $Ψ̃ = (|Ψ̃₁⟩, |Ψ̃₂⟩, … |Ψ̃ₙ⟩, |Ψ⟩)^T$, where $n$ is num_controls. With a matching $G̃$ as in the documentation of TimeDependentGradGenerator, we have

\[G̃ Ψ̃ = \begin{pmatrix} Ĥ |Ψ̃₁⟩ + Ĥ₁|Ψ⟩ \\ \vdots \\ Ĥ |Ψ̃ₙ⟩ + Ĥₙ|Ψ⟩ \\ Ĥ |Ψ⟩ \end{pmatrix}\]

and

\[e^{-i G̃ dt} \begin{pmatrix} 0 \\ \vdots \\ 0 \\ |Ψ⟩ \end{pmatrix} = \begin{pmatrix} \frac{∂}{∂ϵ₁} e^{-i Ĥ dt} |Ψ⟩ \\ \vdots \\ \frac{∂}{∂ϵₙ} e^{-i Ĥ dt} |Ψ⟩ \\ e^{-i Ĥ dt} |Ψ⟩ \end{pmatrix}.\]

QuantumControlBase.initobjpropwrk — Function

wrk = initobjpropwrk(obj, tlist, method; kwargs...)

initializes a workspace for the propagation of an AbstractControlObjective.

QuantumControlBase.getcontrolderivs — Function

Get a vector of the derivatives of generator w.r.t. each control.

getcontrolderivs(generator, controls)

return as vector containing the derivative of generator with respect to each control in controls. The elements of the vector are either nothing if generator does not depend on that particular control, or a function μ(α) that evaluates the derivative for a particular value of the control, see getcontrolderiv.

QuantumControlBase.ConditionalThreads

Private

QuantumControlBase.ConditionalThreads.@threadsif — Macro

Conditionally apply multi-threading to for loops.

This is a variation on Base.Threads.@threads that adds a run-time boolean flag to enable or disable threading. It is intended for internal use in packages building on QuantumControlBase.

Usage:

using QuantumControlBase.ConditionalThreads: @threadsif

function optimize(objectives; use_threads=true)
    @threadsif use_threads for k = 1:length(objectives)
    # ...
    end
end

QuantumControlBase.Functionals

Public

QuantumControlBase.Functionals.F_re — Function

Real-part fidelity.

F_re(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.F_sm — Function

Square-modulus fidelity.

F_sm(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.F_ss — Function

State-to-state phase-insensitive fidelity.

F_ss(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.J_T_re — Function

Real-part functional.

J_T_re(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.J_T_sm — Function

Square-modulus functional.

J_T_sm(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.J_T_ss — Function

State-to-state phase-insensitive functional.

J_T_ss(ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.chi_re! — Function

Krotov-states χ for functional J_T_re.

chi_re!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.chi_sm! — Function

Krotov-states χ for functional J_T_sm.

chi_sm!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.chi_ss! — Function

Krotov-states χ for functional J_T_ss.

chi_ss!(χ, ϕ, objectives; τ=nothing)

QuantumControlBase.Functionals.grad_J_T_sm! — Function

Gradient for J_T_sm.

Private

QuantumControlBase.Functionals.f_tau — Function

Average complex overlap of the target states with forward-propagated states.

f_tau(ϕ, objectives; τ=nothing)

QuantumControlBase.Shapes

Public

QuantumControlBase.Shapes.blackman — Function

Blackman window shape.

blackman(t, t₀, T; a=0.16)

calculates

\[B(t; t_0, T) = \frac{1}{2}\left( 1 - a - \cos\left(2π \frac{t - t_0}{T - t_0}\right) + a \cos\left(4π \frac{t - t_0}{T - t_0}\right) \right)\,,\]

for a scalar t, with $a$ = 0.16.

See http://en.wikipedia.org/wiki/Window_function#Blackman_windows

A Blackman shape looks nearly identical to a Gaussian with a 6-sigma interval between t₀ and T. Unlike the Gaussian, however, it will go exactly to zero at the edges. Thus, Blackman pulses are often preferable to Gaussians.

QuantumControlBase.Shapes.box — Function

Box shape (Theta-function).

box(t, t₀, T)

evaluates the Heaviside (Theta-) function $\Theta(t) = 1$ for $t_0 \le t \le T$; and $\Theta(t) = 0$ otherwise.

QuantumControlBase.Shapes.flattop — Function

Flat shape (one) with a switch-on/switch-off from zero.

flattop(t; t₀, T, t_rise, t_fall=t_rise, func=:blackman)

evaluates a shape function that starts at 0 at $t=t₀$, and ramps to to 1 during the t_rise interval. The function then remains at value 1, before ramping down to 0 again during the interval t_fall before T. For $t < t₀$ and $t > T$, the shape is zero.

The default switch-on/-off shape is half of a Blackman window (see blackman).

For func=:sinsq, the switch-on/-off shape is a sine-squared curve.

QuantumControlBase.TestUtils

Public

QuantumControlBase.TestUtils.random_complex_matrix — Function

Construct a random complex matrix of size N×N with spectral radius ρ.

random_complex_matrix(N, ρ)

QuantumControlBase.TestUtils.random_complex_sparse_matrix — Function

Construct a random sparse complex matrix.

random_complex_sparse_matrix(N, ρ, sparsity)

returns a matrix of size N×N with spectral radius ρ and the given sparsity (number between zero and one that is the approximate fraction of non-zero elements).

QuantumControlBase.TestUtils.random_hermitian_matrix — Function

Construct a random Hermitian matrix of size N×N with spectral radius ρ.

random_hermitian_matrix(N, ρ)

QuantumControlBase.TestUtils.random_hermitian_sparse_matrix — Function

Construct a random sparse Hermitian matrix.

random_hermitian_sparse_matrix(N, ρ, sparsity)

returns a matrix of size N×N with spectral radius ρ and the given sparsity (number between zero and one that is the approximate fraction of non-zero elements).

QuantumControlBase.TestUtils.random_real_matrix — Function

Construct a random real-valued matrix of size N×N with spectral radius ρ.

random_real_matrix(N, ρ)

QuantumControlBase.TestUtils.random_real_sparse_matrix — Function

Construct a random sparse real-valued matrix.

random_real_sparse_matrix(N, ρ, sparsity)

returns a matrix of size N×N with spectral radius ρ and the given sparsity (number between zero and one that is the approximate fraction of non-zero elements).

QuantumControlBase.TestUtils.random_state_vector — Function

Return a random, normalized Hilbert space state vector of dimension N.

random_state_vector(N)